🗺️ Atlas: Extract standard library natives from interpreter blob
