+++ crates/duke-interpreter/src/stdlib.rs
@@ -1,0 +1,5670 @@
+//! Standard library native implementations and bootstrapping.
+
+
+use std::io::Write;
+use duke_runtime::{Slot, VmError, VmResult};
+use crate::{ClassContext, ClassRegistry, NativeControl, NativeThreadAction, FieldEntry, InvokeFn};
+use crate::{format_java_double, format_java_float};
+
+/// Registry of loaded classes — maps class name to its `ClassContext`.
+///
+/// Used by `execute_class` for cross-class method dispatch.
+///
+/// # Examples
+///
+/// A native handler that can call back into the interpreter to invoke Java methods.
+///
+/// The `invoke` closure takes `heap` and `output` as *parameters* (not captured),
+/// using the "loan" pattern: the handler passes its borrows through each call and
+/// gets them back when the call returns. Sequential reborrows — no unsafe required.
+/// Bootstrap minimal JDK standard library classes for native method support.
+///
+/// Creates synthetic `java/lang/System` and `java/io/PrintStream` classes and
+/// registers native `println` handlers for `(Ljava/lang/String;)V`, `(I)V`,
+/// and `()V`.
+#[allow(clippy::too_many_lines)]
+pub fn bootstrap_stdlib(registry: &mut ClassRegistry, heap: &mut duke_gc::Heap) {
+    // Allocate a PrintStream object on the heap.
+    let ps_ref = heap.allocate("java/io/PrintStream".to_string(), 0);
+
+    // Create java/lang/System ClassContext with a single static field `out`.
+    let system_ctx = ClassContext {
+        class_name: "java/lang/System".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: vec![FieldEntry {
+            name: "out".to_string(),
+            descriptor: "Ljava/io/PrintStream;".to_string(),
+            is_static: true,
+        }],
+        static_fields: vec![Slot::Reference(Some(ps_ref))],
+        instance_field_count: 0,
+        interfaces: Vec::new(),
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(system_ctx);
+
+    // Create java/io/PrintStream ClassContext (empty — all methods are native).
+    let ps_ctx = ClassContext {
+        class_name: "java/io/PrintStream".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: Vec::new(),
+        static_fields: Vec::new(),
+        instance_field_count: 0,
+        interfaces: Vec::new(),
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(ps_ctx);
+
+    // Register native println handlers.
+    registry.natives_mut().register(
+        "java/io/PrintStream",
+        "println",
+        "(Ljava/lang/String;)V",
+        native_println_string,
+    );
+    registry
+        .natives_mut()
+        .register("java/io/PrintStream", "println", "(I)V", native_println_int);
+    registry
+        .natives_mut()
+        .register("java/io/PrintStream", "println", "()V", native_println_void);
+
+    let file_ctx = ClassContext {
+        class_name: "java/io/File".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: vec![FieldEntry {
+            name: "path".to_string(),
+            descriptor: "Ljava/lang/String;".to_string(),
+            is_static: false,
+        }],
+        static_fields: Vec::new(),
+        instance_field_count: 1,
+        interfaces: Vec::new(),
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(file_ctx);
+    registry.natives_mut().register(
+        "java/io/File",
+        "<init>",
+        "(Ljava/lang/String;)V",
+        native_file_init,
+    );
+    registry
+        .natives_mut()
+        .register("java/io/File", "exists", "()Z", native_file_exists);
+    registry
+        .natives_mut()
+        .register("java/io/File", "isFile", "()Z", native_file_is_file);
+    registry.natives_mut().register(
+        "java/io/File",
+        "isDirectory",
+        "()Z",
+        native_file_is_directory,
+    );
+
+    let io_exception_ctx = ClassContext {
+        class_name: "java/io/IOException".to_string(),
+        super_class: Some("java/lang/Exception".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: Vec::new(),
+        static_fields: Vec::new(),
+        instance_field_count: 0,
+        interfaces: Vec::new(),
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(io_exception_ctx);
+
+    let file_not_found_ctx = ClassContext {
+        class_name: "java/io/FileNotFoundException".to_string(),
+        super_class: Some("java/io/IOException".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: Vec::new(),
+        static_fields: Vec::new(),
+        instance_field_count: 0,
+        interfaces: Vec::new(),
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(file_not_found_ctx);
+
+    let socket_exception_ctx = ClassContext {
+        class_name: "java/net/SocketException".to_string(),
+        super_class: Some("java/io/IOException".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: Vec::new(),
+        static_fields: Vec::new(),
+        instance_field_count: 0,
+        interfaces: Vec::new(),
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(socket_exception_ctx);
+
+    let bind_exception_ctx = ClassContext {
+        class_name: "java/net/BindException".to_string(),
+        super_class: Some("java/net/SocketException".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: Vec::new(),
+        static_fields: Vec::new(),
+        instance_field_count: 0,
+        interfaces: Vec::new(),
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(bind_exception_ctx);
+
+    let connect_exception_ctx = ClassContext {
+        class_name: "java/net/ConnectException".to_string(),
+        super_class: Some("java/net/SocketException".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: Vec::new(),
+        static_fields: Vec::new(),
+        instance_field_count: 0,
+        interfaces: Vec::new(),
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(connect_exception_ctx);
+
+    let input_stream_ctx = ClassContext {
+        class_name: "java/io/InputStream".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: Vec::new(),
+        static_fields: Vec::new(),
+        instance_field_count: 0,
+        interfaces: vec!["java/lang/AutoCloseable".to_string()],
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(input_stream_ctx);
+
+    let output_stream_ctx = ClassContext {
+        class_name: "java/io/OutputStream".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: Vec::new(),
+        static_fields: Vec::new(),
+        instance_field_count: 0,
+        interfaces: vec!["java/lang/AutoCloseable".to_string()],
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(output_stream_ctx);
+
+    let file_input_stream_ctx = ClassContext {
+        class_name: "java/io/FileInputStream".to_string(),
+        super_class: Some("java/io/InputStream".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: vec![FieldEntry {
+            name: "fd".to_string(),
+            descriptor: "I".to_string(),
+            is_static: false,
+        }],
+        static_fields: Vec::new(),
+        instance_field_count: 1,
+        interfaces: vec!["java/lang/AutoCloseable".to_string()],
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(file_input_stream_ctx);
+    registry.natives_mut().register(
+        "java/io/FileInputStream",
+        "<init>",
+        "(Ljava/lang/String;)V",
+        native_file_input_stream_init,
+    );
+    registry.natives_mut().register(
+        "java/io/FileInputStream",
+        "read",
+        "()I",
+        native_file_input_stream_read,
+    );
+    registry.natives_mut().register(
+        "java/io/FileInputStream",
+        "read",
+        "([B)I",
+        native_file_input_stream_read_bytes,
+    );
+    registry.natives_mut().register(
+        "java/io/FileInputStream",
+        "close",
+        "()V",
+        native_file_input_stream_close,
+    );
+
+    let file_output_stream_ctx = ClassContext {
+        class_name: "java/io/FileOutputStream".to_string(),
+        super_class: Some("java/io/OutputStream".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: vec![FieldEntry {
+            name: "fd".to_string(),
+            descriptor: "I".to_string(),
+            is_static: false,
+        }],
+        static_fields: Vec::new(),
+        instance_field_count: 1,
+        interfaces: vec!["java/lang/AutoCloseable".to_string()],
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(file_output_stream_ctx);
+    registry.natives_mut().register(
+        "java/io/FileOutputStream",
+        "<init>",
+        "(Ljava/lang/String;)V",
+        native_file_output_stream_init,
+    );
+    registry.natives_mut().register(
+        "java/io/FileOutputStream",
+        "write",
+        "(I)V",
+        native_file_output_stream_write,
+    );
+    registry.natives_mut().register(
+        "java/io/FileOutputStream",
+        "write",
+        "([B)V",
+        native_file_output_stream_write_bytes,
+    );
+    registry.natives_mut().register(
+        "java/io/FileOutputStream",
+        "close",
+        "()V",
+        native_file_output_stream_close,
+    );
+
+    let server_socket_ctx = ClassContext {
+        class_name: "java/net/ServerSocket".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: vec![
+            FieldEntry {
+                name: "fd".to_string(),
+                descriptor: "I".to_string(),
+                is_static: false,
+            },
+            FieldEntry {
+                name: "port".to_string(),
+                descriptor: "I".to_string(),
+                is_static: false,
+            },
+        ],
+        static_fields: Vec::new(),
+        instance_field_count: 2,
+        interfaces: vec!["java/lang/AutoCloseable".to_string()],
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(server_socket_ctx);
+    registry.natives_mut().register(
+        "java/net/ServerSocket",
+        "<init>",
+        "(I)V",
+        native_server_socket_init,
+    );
+    registry.natives_mut().register(
+        "java/net/ServerSocket",
+        "accept",
+        "()Ljava/net/Socket;",
+        native_server_socket_accept,
+    );
+    registry.natives_mut().register(
+        "java/net/ServerSocket",
+        "getLocalPort",
+        "()I",
+        native_server_socket_get_local_port,
+    );
+    registry.natives_mut().register(
+        "java/net/ServerSocket",
+        "close",
+        "()V",
+        native_server_socket_close,
+    );
+
+    let socket_ctx = ClassContext {
+        class_name: "java/net/Socket".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: vec![
+            FieldEntry {
+                name: "fdRead".to_string(),
+                descriptor: "I".to_string(),
+                is_static: false,
+            },
+            FieldEntry {
+                name: "fdWrite".to_string(),
+                descriptor: "I".to_string(),
+                is_static: false,
+            },
+        ],
+        static_fields: Vec::new(),
+        instance_field_count: 2,
+        interfaces: vec!["java/lang/AutoCloseable".to_string()],
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(socket_ctx);
+    registry.natives_mut().register(
+        "java/net/Socket",
+        "<init>",
+        "(Ljava/lang/String;I)V",
+        native_socket_init,
+    );
+    registry.natives_mut().register(
+        "java/net/Socket",
+        "getInputStream",
+        "()Ljava/io/InputStream;",
+        native_socket_get_input_stream,
+    );
+    registry.natives_mut().register(
+        "java/net/Socket",
+        "getOutputStream",
+        "()Ljava/io/OutputStream;",
+        native_socket_get_output_stream,
+    );
+    registry
+        .natives_mut()
+        .register("java/net/Socket", "close", "()V", native_socket_close);
+
+    let socket_input_stream_ctx = ClassContext {
+        class_name: "duke/net/SocketInputStream".to_string(),
+        super_class: Some("java/io/InputStream".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: vec![FieldEntry {
+            name: "fd".to_string(),
+            descriptor: "I".to_string(),
+            is_static: false,
+        }],
+        static_fields: Vec::new(),
+        instance_field_count: 1,
+        interfaces: vec!["java/lang/AutoCloseable".to_string()],
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(socket_input_stream_ctx);
+    registry.natives_mut().register(
+        "duke/net/SocketInputStream",
+        "read",
+        "()I",
+        native_file_input_stream_read,
+    );
+    registry.natives_mut().register(
+        "duke/net/SocketInputStream",
+        "read",
+        "([B)I",
+        native_file_input_stream_read_bytes,
+    );
+    registry.natives_mut().register(
+        "duke/net/SocketInputStream",
+        "close",
+        "()V",
+        native_file_input_stream_close,
+    );
+
+    let socket_output_stream_ctx = ClassContext {
+        class_name: "duke/net/SocketOutputStream".to_string(),
+        super_class: Some("java/io/OutputStream".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: vec![FieldEntry {
+            name: "fd".to_string(),
+            descriptor: "I".to_string(),
+            is_static: false,
+        }],
+        static_fields: Vec::new(),
+        instance_field_count: 1,
+        interfaces: vec!["java/lang/AutoCloseable".to_string()],
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(socket_output_stream_ctx);
+    registry.natives_mut().register(
+        "duke/net/SocketOutputStream",
+        "write",
+        "(I)V",
+        native_file_output_stream_write,
+    );
+    registry.natives_mut().register(
+        "duke/net/SocketOutputStream",
+        "write",
+        "([B)V",
+        native_file_output_stream_write_bytes,
+    );
+    registry.natives_mut().register(
+        "duke/net/SocketOutputStream",
+        "close",
+        "()V",
+        native_file_output_stream_close,
+    );
+
+    // Register java/lang/String ClassContext (empty — instance methods are native).
+    let string_ctx = ClassContext {
+        class_name: "java/lang/String".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: Vec::new(),
+        static_fields: Vec::new(),
+        instance_field_count: 0,
+        interfaces: vec![
+            "java/lang/Comparable".to_string(),
+            "java/io/Serializable".to_string(),
+            "java/lang/CharSequence".to_string(),
+        ],
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(string_ctx);
+
+    // String instance methods
+    registry
+        .natives_mut()
+        .register("java/lang/String", "length", "()I", native_string_length);
+    registry.natives_mut().register(
+        "java/lang/String",
+        "equals",
+        "(Ljava/lang/Object;)Z",
+        native_string_equals,
+    );
+    registry
+        .natives_mut()
+        .register("java/lang/String", "charAt", "(I)C", native_string_char_at);
+    registry.natives_mut().register(
+        "java/lang/String",
+        "substring",
+        "(I)Ljava/lang/String;",
+        native_string_substring,
+    );
+    registry.natives_mut().register(
+        "java/lang/String",
+        "substring",
+        "(II)Ljava/lang/String;",
+        native_string_substring_range,
+    );
+    registry.natives_mut().register(
+        "java/lang/String",
+        "indexOf",
+        "(Ljava/lang/String;)I",
+        native_string_indexof,
+    );
+    registry.natives_mut().register(
+        "java/lang/String",
+        "contains",
+        "(Ljava/lang/CharSequence;)Z",
+        native_string_contains,
+    );
+    registry
+        .natives_mut()
+        .register("java/lang/String", "isEmpty", "()Z", native_string_isempty);
+    registry.natives_mut().register(
+        "java/lang/String",
+        "compareTo",
+        "(Ljava/lang/String;)I",
+        native_string_compareto,
+    );
+    registry.natives_mut().register(
+        "java/lang/String",
+        "compareTo",
+        "(Ljava/lang/Object;)I",
+        native_string_compareto_object,
+    );
+    registry.natives_mut().register(
+        "java/lang/String",
+        "startsWith",
+        "(Ljava/lang/String;)Z",
+        native_string_startswith,
+    );
+    registry.natives_mut().register(
+        "java/lang/String",
+        "endsWith",
+        "(Ljava/lang/String;)Z",
+        native_string_endswith,
+    );
+    registry.natives_mut().register(
+        "java/lang/String",
+        "trim",
+        "()Ljava/lang/String;",
+        native_string_trim,
+    );
+    registry.natives_mut().register(
+        "java/lang/String",
+        "toCharArray",
+        "()[C",
+        native_string_tochararray,
+    );
+    registry.natives_mut().register(
+        "java/lang/String",
+        "toUpperCase",
+        "()Ljava/lang/String;",
+        native_string_touppercase,
+    );
+    registry.natives_mut().register(
+        "java/lang/String",
+        "toLowerCase",
+        "()Ljava/lang/String;",
+        native_string_tolowercase,
+    );
+    registry.natives_mut().register(
+        "java/lang/String",
+        "replace",
+        "(CC)Ljava/lang/String;",
+        native_string_replace_char,
+    );
+    registry.natives_mut().register(
+        "java/lang/String",
+        "replace",
+        "(Ljava/lang/CharSequence;Ljava/lang/CharSequence;)Ljava/lang/String;",
+        native_string_replace_charsequence,
+    );
+    registry.natives_mut().register(
+        "java/lang/String",
+        "split",
+        "(Ljava/lang/String;)[Ljava/lang/String;",
+        native_string_split,
+    );
+    registry.natives_mut().register(
+        "java/lang/String",
+        "hashCode",
+        "()I",
+        native_string_hashcode,
+    );
+    registry.natives_mut().register(
+        "java/lang/String",
+        "toString",
+        "()Ljava/lang/String;",
+        native_string_tostring,
+    );
+
+    // String static methods
+    registry.natives_mut().register(
+        "java/lang/String",
+        "valueOf",
+        "(I)Ljava/lang/String;",
+        native_string_value_of_int,
+    );
+
+    // PrintStream.print (no newline)
+    registry.natives_mut().register(
+        "java/io/PrintStream",
+        "print",
+        "(Ljava/lang/String;)V",
+        native_print_string,
+    );
+    registry
+        .natives_mut()
+        .register("java/io/PrintStream", "print", "(I)V", native_print_int);
+
+    // println overloads
+    registry.natives_mut().register(
+        "java/io/PrintStream",
+        "println",
+        "(J)V",
+        native_println_long,
+    );
+    registry.natives_mut().register(
+        "java/io/PrintStream",
+        "println",
+        "(F)V",
+        native_println_float,
+    );
+    registry.natives_mut().register(
+        "java/io/PrintStream",
+        "println",
+        "(D)V",
+        native_println_double,
+    );
+    registry.natives_mut().register(
+        "java/io/PrintStream",
+        "println",
+        "(Z)V",
+        native_println_boolean,
+    );
+    registry.natives_mut().register(
+        "java/io/PrintStream",
+        "println",
+        "(C)V",
+        native_println_char,
+    );
+    registry.natives_mut().register(
+        "java/io/PrintStream",
+        "println",
+        "(Ljava/lang/Object;)V",
+        native_println_object,
+    );
+
+    // print overloads
+    registry
+        .natives_mut()
+        .register("java/io/PrintStream", "print", "(J)V", native_print_long);
+    registry
+        .natives_mut()
+        .register("java/io/PrintStream", "print", "(F)V", native_print_float);
+    registry
+        .natives_mut()
+        .register("java/io/PrintStream", "print", "(D)V", native_print_double);
+    registry
+        .natives_mut()
+        .register("java/io/PrintStream", "print", "(Z)V", native_print_boolean);
+    registry
+        .natives_mut()
+        .register("java/io/PrintStream", "print", "(C)V", native_print_char);
+    registry.natives_mut().register(
+        "java/io/PrintStream",
+        "print",
+        "(Ljava/lang/Object;)V",
+        native_print_object,
+    );
+
+    // System.exit (static)
+    registry
+        .natives_mut()
+        .register("java/lang/System", "exit", "(I)V", native_system_exit);
+
+    // Register synthetic exception hierarchy so is_assignable_from can walk it.
+    // java/lang/Object (root — no super)
+    let object_ctx = ClassContext {
+        class_name: "java/lang/Object".to_string(),
+        super_class: None,
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: Vec::new(),
+        static_fields: Vec::new(),
+        instance_field_count: 0,
+        interfaces: Vec::new(),
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(object_ctx);
+
+    // Object instance methods
+    registry.natives_mut().register(
+        "java/lang/Object",
+        "hashCode",
+        "()I",
+        native_object_hashcode,
+    );
+    registry.natives_mut().register(
+        "java/lang/Object",
+        "toString",
+        "()Ljava/lang/String;",
+        native_object_tostring,
+    );
+    registry.natives_mut().register(
+        "java/lang/Object",
+        "clone",
+        "()Ljava/lang/Object;",
+        native_object_clone,
+    );
+
+    // java/lang/Class — lightweight stub for class literals
+    let class_ctx = ClassContext {
+        class_name: "java/lang/Class".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: Vec::new(),
+        static_fields: Vec::new(),
+        instance_field_count: 0,
+        interfaces: Vec::new(),
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(class_ctx);
+
+    let runnable_ctx = ClassContext {
+        class_name: "java/lang/Runnable".to_string(),
+        super_class: None,
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: Vec::new(),
+        static_fields: Vec::new(),
+        instance_field_count: 0,
+        interfaces: Vec::new(),
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(runnable_ctx);
+
+    let thread_ctx = ClassContext {
+        class_name: "java/lang/Thread".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: vec![
+            FieldEntry {
+                name: "target".to_string(),
+                descriptor: "Ljava/lang/Runnable;".to_string(),
+                is_static: false,
+            },
+            FieldEntry {
+                name: "threadId".to_string(),
+                descriptor: "I".to_string(),
+                is_static: false,
+            },
+        ],
+        static_fields: Vec::new(),
+        instance_field_count: 2,
+        interfaces: Vec::new(),
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(thread_ctx);
+    registry
+        .natives_mut()
+        .register("java/lang/Thread", "<init>", "()V", native_thread_init);
+    registry.natives_mut().register(
+        "java/lang/Thread",
+        "<init>",
+        "(Ljava/lang/Runnable;)V",
+        native_thread_init_runnable,
+    );
+    registry
+        .natives_mut()
+        .register("java/lang/Thread", "start", "()V", native_thread_start);
+    registry
+        .natives_mut()
+        .register("java/lang/Thread", "join", "()V", native_thread_join);
+    registry
+        .natives_mut()
+        .register("java/lang/Thread", "sleep", "(J)V", native_thread_sleep);
+
+    // java/lang/Throwable extends Object
+    let throwable_ctx = ClassContext {
+        class_name: "java/lang/Throwable".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: Vec::new(),
+        static_fields: Vec::new(),
+        instance_field_count: 0,
+        interfaces: Vec::new(),
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(throwable_ctx);
+    registry.natives_mut().register(
+        "java/lang/Throwable",
+        "addSuppressed",
+        "(Ljava/lang/Throwable;)V",
+        native_throwable_add_suppressed,
+    );
+
+    // java/lang/Exception extends Throwable
+    let exception_ctx = ClassContext {
+        class_name: "java/lang/Exception".to_string(),
+        super_class: Some("java/lang/Throwable".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: Vec::new(),
+        static_fields: Vec::new(),
+        instance_field_count: 0,
+        interfaces: Vec::new(),
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(exception_ctx);
+
+    // java/lang/RuntimeException extends Exception
+    let rte_ctx = ClassContext {
+        class_name: "java/lang/RuntimeException".to_string(),
+        super_class: Some("java/lang/Exception".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: Vec::new(),
+        static_fields: Vec::new(),
+        instance_field_count: 0,
+        interfaces: Vec::new(),
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(rte_ctx);
+
+    // java/lang/AutoCloseable — marker interface for try-with-resources.
+    // Registered so is_assignable_from correctly handles queries like
+    // "does TryWithResources$Res implement AutoCloseable?" without
+    // erroring on an unknown class.
+    let autocloseable_ctx = ClassContext {
+        class_name: "java/lang/AutoCloseable".to_string(),
+        super_class: None, // interface — no super class
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: Vec::new(),
+        static_fields: Vec::new(),
+        instance_field_count: 0,
+        interfaces: Vec::new(),
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(autocloseable_ctx);
+
+    // java/lang/Enum — abstract superclass for all enums.
+    // Fields: name (String) at index 0, ordinal (int) at index 1.
+    let enum_ctx = ClassContext {
+        class_name: "java/lang/Enum".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: vec![
+            FieldEntry {
+                name: "name".to_string(),
+                descriptor: "Ljava/lang/String;".to_string(),
+                is_static: false,
+            },
+            FieldEntry {
+                name: "ordinal".to_string(),
+                descriptor: "I".to_string(),
+                is_static: false,
+            },
+        ],
+        static_fields: Vec::new(),
+        instance_field_count: 2,
+        interfaces: Vec::new(),
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(enum_ctx);
+    registry.natives_mut().register(
+        "java/lang/Enum",
+        "<init>",
+        "(Ljava/lang/String;I)V",
+        native_enum_init,
+    );
+    registry
+        .natives_mut()
+        .register("java/lang/Enum", "ordinal", "()I", native_enum_ordinal);
+    registry.natives_mut().register(
+        "java/lang/Enum",
+        "name",
+        "()Ljava/lang/String;",
+        native_enum_name,
+    );
+    registry.natives_mut().register(
+        "java/lang/Enum",
+        "valueOf",
+        "(Ljava/lang/Class;Ljava/lang/String;)Ljava/lang/Enum;",
+        native_enum_valueof,
+    );
+
+    // java/lang/Integer — boxed int with value field + numeric constants
+    let integer_ctx = ClassContext {
+        class_name: "java/lang/Integer".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: vec![
+            FieldEntry {
+                name: "value".to_string(),
+                descriptor: "I".to_string(),
+                is_static: false,
+            },
+            FieldEntry {
+                name: "MAX_VALUE".to_string(),
+                descriptor: "I".to_string(),
+                is_static: true,
+            },
+            FieldEntry {
+                name: "MIN_VALUE".to_string(),
+                descriptor: "I".to_string(),
+                is_static: true,
+            },
+        ],
+        static_fields: vec![Slot::Int(i32::MAX), Slot::Int(i32::MIN)],
+        instance_field_count: 1,
+        interfaces: vec!["java/lang/Comparable".to_string()],
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(integer_ctx);
+    registry.natives_mut().register(
+        "java/lang/Integer",
+        "parseInt",
+        "(Ljava/lang/String;)I",
+        native_integer_parseint,
+    );
+    registry.natives_mut().register(
+        "java/lang/Integer",
+        "valueOf",
+        "(I)Ljava/lang/Integer;",
+        native_integer_valueof,
+    );
+    registry.natives_mut().register(
+        "java/lang/Integer",
+        "intValue",
+        "()I",
+        native_integer_intvalue,
+    );
+    registry.natives_mut().register(
+        "java/lang/Integer",
+        "toString",
+        "(I)Ljava/lang/String;",
+        native_integer_tostring_static,
+    );
+    registry.natives_mut().register(
+        "java/lang/Integer",
+        "compareTo",
+        "(Ljava/lang/Object;)I",
+        native_integer_compareto,
+    );
+
+    // java/lang/Long — boxed long with value field + numeric constants
+    let long_ctx = ClassContext {
+        class_name: "java/lang/Long".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: vec![
+            FieldEntry {
+                name: "value".to_string(),
+                descriptor: "J".to_string(),
+                is_static: false,
+            },
+            FieldEntry {
+                name: "MAX_VALUE".to_string(),
+                descriptor: "J".to_string(),
+                is_static: true,
+            },
+            FieldEntry {
+                name: "MIN_VALUE".to_string(),
+                descriptor: "J".to_string(),
+                is_static: true,
+            },
+        ],
+        static_fields: vec![Slot::Long(i64::MAX), Slot::Long(i64::MIN)],
+        instance_field_count: 1,
+        interfaces: vec!["java/lang/Comparable".to_string()],
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(long_ctx);
+    registry.natives_mut().register(
+        "java/lang/Long",
+        "parseLong",
+        "(Ljava/lang/String;)J",
+        native_long_parselong,
+    );
+    registry.natives_mut().register(
+        "java/lang/Long",
+        "valueOf",
+        "(J)Ljava/lang/Long;",
+        native_long_valueof,
+    );
+    registry
+        .natives_mut()
+        .register("java/lang/Long", "longValue", "()J", native_long_longvalue);
+    registry.natives_mut().register(
+        "java/lang/Long",
+        "toString",
+        "(J)Ljava/lang/String;",
+        native_long_tostring_static,
+    );
+    registry.natives_mut().register(
+        "java/lang/Long",
+        "compareTo",
+        "(Ljava/lang/Object;)I",
+        native_long_compareto,
+    );
+
+    // java/lang/Double — boxed double with value field + numeric constants
+    let double_ctx = ClassContext {
+        class_name: "java/lang/Double".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: vec![
+            FieldEntry {
+                name: "value".to_string(),
+                descriptor: "D".to_string(),
+                is_static: false,
+            },
+            FieldEntry {
+                name: "MAX_VALUE".to_string(),
+                descriptor: "D".to_string(),
+                is_static: true,
+            },
+            FieldEntry {
+                name: "MIN_VALUE".to_string(),
+                descriptor: "D".to_string(),
+                is_static: true,
+            },
+            FieldEntry {
+                name: "NaN".to_string(),
+                descriptor: "D".to_string(),
+                is_static: true,
+            },
+            FieldEntry {
+                name: "POSITIVE_INFINITY".to_string(),
+                descriptor: "D".to_string(),
+                is_static: true,
+            },
+            FieldEntry {
+                name: "NEGATIVE_INFINITY".to_string(),
+                descriptor: "D".to_string(),
+                is_static: true,
+            },
+        ],
+        static_fields: vec![
+            Slot::Double(f64::MAX),
+            Slot::Double(5e-324_f64),
+            Slot::Double(f64::NAN),
+            Slot::Double(f64::INFINITY),
+            Slot::Double(f64::NEG_INFINITY),
+        ],
+        instance_field_count: 1,
+        interfaces: vec!["java/lang/Comparable".to_string()],
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(double_ctx);
+    registry.natives_mut().register(
+        "java/lang/Double",
+        "parseDouble",
+        "(Ljava/lang/String;)D",
+        native_double_parsedouble,
+    );
+    registry.natives_mut().register(
+        "java/lang/Double",
+        "valueOf",
+        "(D)Ljava/lang/Double;",
+        native_double_valueof,
+    );
+    registry.natives_mut().register(
+        "java/lang/Double",
+        "doubleValue",
+        "()D",
+        native_double_doublevalue,
+    );
+    registry
+        .natives_mut()
+        .register("java/lang/Double", "isNaN", "(D)Z", native_double_isnan);
+    registry.natives_mut().register(
+        "java/lang/Double",
+        "compareTo",
+        "(Ljava/lang/Object;)I",
+        native_double_compareto,
+    );
+
+    // java/lang/Float — boxed float with value field
+    let float_ctx = ClassContext {
+        class_name: "java/lang/Float".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: vec![FieldEntry {
+            name: "value".to_string(),
+            descriptor: "F".to_string(),
+            is_static: false,
+        }],
+        static_fields: Vec::new(),
+        instance_field_count: 1,
+        interfaces: vec!["java/lang/Comparable".to_string()],
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(float_ctx);
+    registry.natives_mut().register(
+        "java/lang/Float",
+        "parseFloat",
+        "(Ljava/lang/String;)F",
+        native_float_parsefloat,
+    );
+
+    // java/lang/Boolean — static utility
+    let boolean_ctx = ClassContext {
+        class_name: "java/lang/Boolean".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: Vec::new(),
+        static_fields: Vec::new(),
+        instance_field_count: 0,
+        interfaces: vec!["java/lang/Comparable".to_string()],
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(boolean_ctx);
+    registry.natives_mut().register(
+        "java/lang/Boolean",
+        "parseBoolean",
+        "(Ljava/lang/String;)Z",
+        native_boolean_parseboolean,
+    );
+
+    // java/lang/Math — static math utilities with PI and E constants
+    let math_ctx = ClassContext {
+        class_name: "java/lang/Math".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: vec![
+            FieldEntry {
+                name: "PI".to_string(),
+                descriptor: "D".to_string(),
+                is_static: true,
+            },
+            FieldEntry {
+                name: "E".to_string(),
+                descriptor: "D".to_string(),
+                is_static: true,
+            },
+        ],
+        static_fields: vec![
+            Slot::Double(std::f64::consts::PI),
+            Slot::Double(std::f64::consts::E),
+        ],
+        instance_field_count: 0,
+        interfaces: Vec::new(),
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(math_ctx);
+    registry
+        .natives_mut()
+        .register("java/lang/Math", "max", "(II)I", native_math_max_int);
+    registry
+        .natives_mut()
+        .register("java/lang/Math", "min", "(II)I", native_math_min_int);
+    registry
+        .natives_mut()
+        .register("java/lang/Math", "abs", "(I)I", native_math_abs_int);
+    registry
+        .natives_mut()
+        .register("java/lang/Math", "sqrt", "(D)D", native_math_sqrt);
+    registry
+        .natives_mut()
+        .register("java/lang/Math", "pow", "(DD)D", native_math_pow);
+    registry
+        .natives_mut()
+        .register("java/lang/Math", "floor", "(D)D", native_math_floor);
+    registry
+        .natives_mut()
+        .register("java/lang/Math", "ceil", "(D)D", native_math_ceil);
+    registry
+        .natives_mut()
+        .register("java/lang/Math", "round", "(D)J", native_math_round_double);
+    registry
+        .natives_mut()
+        .register("java/lang/Math", "abs", "(J)J", native_math_abs_long);
+    registry
+        .natives_mut()
+        .register("java/lang/Math", "abs", "(D)D", native_math_abs_double);
+    registry
+        .natives_mut()
+        .register("java/lang/Math", "max", "(JJ)J", native_math_max_long);
+    registry
+        .natives_mut()
+        .register("java/lang/Math", "min", "(JJ)J", native_math_min_long);
+    registry
+        .natives_mut()
+        .register("java/lang/Math", "max", "(DD)D", native_math_max_double);
+    registry
+        .natives_mut()
+        .register("java/lang/Math", "min", "(DD)D", native_math_min_double);
+
+    // String.valueOf overloads (int already registered above)
+    registry.natives_mut().register(
+        "java/lang/String",
+        "valueOf",
+        "(J)Ljava/lang/String;",
+        native_string_value_of_long,
+    );
+    registry.natives_mut().register(
+        "java/lang/String",
+        "valueOf",
+        "(D)Ljava/lang/String;",
+        native_string_value_of_double,
+    );
+    registry.natives_mut().register(
+        "java/lang/String",
+        "valueOf",
+        "(F)Ljava/lang/String;",
+        native_string_value_of_float,
+    );
+    registry.natives_mut().register(
+        "java/lang/String",
+        "valueOf",
+        "(Z)Ljava/lang/String;",
+        native_string_value_of_boolean,
+    );
+    registry.natives_mut().register(
+        "java/lang/String",
+        "valueOf",
+        "(C)Ljava/lang/String;",
+        native_string_value_of_char,
+    );
+    registry.natives_mut().register(
+        "java/lang/String",
+        "valueOf",
+        "(Ljava/lang/Object;)Ljava/lang/String;",
+        native_string_value_of_object,
+    );
+
+    // String.concat
+    registry.natives_mut().register(
+        "java/lang/String",
+        "concat",
+        "(Ljava/lang/String;)Ljava/lang/String;",
+        native_string_concat,
+    );
+
+    // String.format(String, Object[]) -> String
+    registry.natives_mut().register(
+        "java/lang/String",
+        "format",
+        "(Ljava/lang/String;[Ljava/lang/Object;)Ljava/lang/String;",
+        native_string_format,
+    );
+
+    // java/lang/StringBuilder — mutable string buffer
+    let sb_ctx = ClassContext {
+        class_name: "java/lang/StringBuilder".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: Vec::new(),
+        static_fields: Vec::new(),
+        instance_field_count: 0,
+        interfaces: vec!["java/lang/CharSequence".to_string()],
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(sb_ctx);
+
+    // StringBuilder.<init>()V
+    registry
+        .natives_mut()
+        .register("java/lang/StringBuilder", "<init>", "()V", native_sb_init);
+    // StringBuilder.<init>(String)V
+    registry.natives_mut().register(
+        "java/lang/StringBuilder",
+        "<init>",
+        "(Ljava/lang/String;)V",
+        native_sb_init_string,
+    );
+    // StringBuilder.append(String)
+    registry.natives_mut().register(
+        "java/lang/StringBuilder",
+        "append",
+        "(Ljava/lang/String;)Ljava/lang/StringBuilder;",
+        native_sb_append_string,
+    );
+    // StringBuilder.append(int)
+    registry.natives_mut().register(
+        "java/lang/StringBuilder",
+        "append",
+        "(I)Ljava/lang/StringBuilder;",
+        native_sb_append_int,
+    );
+    // StringBuilder.append(long)
+    registry.natives_mut().register(
+        "java/lang/StringBuilder",
+        "append",
+        "(J)Ljava/lang/StringBuilder;",
+        native_sb_append_long,
+    );
+    // StringBuilder.append(double)
+    registry.natives_mut().register(
+        "java/lang/StringBuilder",
+        "append",
+        "(D)Ljava/lang/StringBuilder;",
+        native_sb_append_double,
+    );
+    // StringBuilder.append(float)
+    registry.natives_mut().register(
+        "java/lang/StringBuilder",
+        "append",
+        "(F)Ljava/lang/StringBuilder;",
+        native_sb_append_float,
+    );
+    // StringBuilder.append(boolean)
+    registry.natives_mut().register(
+        "java/lang/StringBuilder",
+        "append",
+        "(Z)Ljava/lang/StringBuilder;",
+        native_sb_append_boolean,
+    );
+    // StringBuilder.append(char)
+    registry.natives_mut().register(
+        "java/lang/StringBuilder",
+        "append",
+        "(C)Ljava/lang/StringBuilder;",
+        native_sb_append_char,
+    );
+    // StringBuilder.toString()
+    registry.natives_mut().register(
+        "java/lang/StringBuilder",
+        "toString",
+        "()Ljava/lang/String;",
+        native_sb_tostring,
+    );
+    // StringBuilder.length()
+    registry
+        .natives_mut()
+        .register("java/lang/StringBuilder", "length", "()I", native_sb_length);
+
+    // java/lang/Character — static character utilities + boxed char
+    let character_ctx = ClassContext {
+        class_name: "java/lang/Character".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: vec![FieldEntry {
+            name: "value".to_string(),
+            descriptor: "C".to_string(),
+            is_static: false,
+        }],
+        static_fields: Vec::new(),
+        instance_field_count: 1,
+        interfaces: vec!["java/lang/Comparable".to_string()],
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(character_ctx);
+
+    // Character.isDigit(C)Z
+    registry.natives_mut().register(
+        "java/lang/Character",
+        "isDigit",
+        "(C)Z",
+        native_char_is_digit,
+    );
+    // Character.isLetter(C)Z
+    registry.natives_mut().register(
+        "java/lang/Character",
+        "isLetter",
+        "(C)Z",
+        native_char_is_letter,
+    );
+    // Character.isWhitespace(C)Z
+    registry.natives_mut().register(
+        "java/lang/Character",
+        "isWhitespace",
+        "(C)Z",
+        native_char_is_whitespace,
+    );
+    // Character.isUpperCase(C)Z
+    registry.natives_mut().register(
+        "java/lang/Character",
+        "isUpperCase",
+        "(C)Z",
+        native_char_is_uppercase,
+    );
+    // Character.isLowerCase(C)Z
+    registry.natives_mut().register(
+        "java/lang/Character",
+        "isLowerCase",
+        "(C)Z",
+        native_char_is_lowercase,
+    );
+    // Character.toUpperCase(C)C
+    registry.natives_mut().register(
+        "java/lang/Character",
+        "toUpperCase",
+        "(C)C",
+        native_char_to_uppercase,
+    );
+    // Character.toLowerCase(C)C
+    registry.natives_mut().register(
+        "java/lang/Character",
+        "toLowerCase",
+        "(C)C",
+        native_char_to_lowercase,
+    );
+    // Character.isLetterOrDigit(C)Z
+    registry.natives_mut().register(
+        "java/lang/Character",
+        "isLetterOrDigit",
+        "(C)Z",
+        native_char_is_letter_or_digit,
+    );
+    // Character.valueOf(C)Ljava/lang/Character;
+    registry.natives_mut().register(
+        "java/lang/Character",
+        "valueOf",
+        "(C)Ljava/lang/Character;",
+        native_char_valueof,
+    );
+    // Character.charValue()C
+    registry.natives_mut().register(
+        "java/lang/Character",
+        "charValue",
+        "()C",
+        native_char_charvalue,
+    );
+
+    // java/util/ArrayList — dynamic list backed by growable fields
+    // fields[0] = size (Int), fields[1..] = elements (pushed dynamically)
+    let arraylist_ctx = ClassContext {
+        class_name: "java/util/ArrayList".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: vec![FieldEntry {
+            name: "size".to_string(),
+            descriptor: "I".to_string(),
+            is_static: false,
+        }],
+        static_fields: Vec::new(),
+        instance_field_count: 1,
+        interfaces: vec![
+            "java/util/List".to_string(),
+            "java/util/Collection".to_string(),
+            "java/lang/Iterable".to_string(),
+        ],
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(arraylist_ctx);
+    registry.natives_mut().register(
+        "java/util/ArrayList",
+        "<init>",
+        "()V",
+        native_arraylist_init,
+    );
+    registry.natives_mut().register(
+        "java/util/ArrayList",
+        "add",
+        "(Ljava/lang/Object;)Z",
+        native_arraylist_add,
+    );
+    registry.natives_mut().register(
+        "java/util/ArrayList",
+        "get",
+        "(I)Ljava/lang/Object;",
+        native_arraylist_get,
+    );
+    registry
+        .natives_mut()
+        .register("java/util/ArrayList", "size", "()I", native_arraylist_size);
+    registry.natives_mut().register(
+        "java/util/ArrayList",
+        "iterator",
+        "()Ljava/util/Iterator;",
+        native_arraylist_iterator,
+    );
+    registry.natives_mut().register_callback(
+        "java/util/ArrayList",
+        "sort",
+        "(Ljava/util/Comparator;)V",
+        array_list_sort,
+    );
+
+    // duke/util/ArrayListIterator — internal iterator for ArrayList
+    // fields[0] = ArrayList reference, fields[1] = current index (Int)
+    let iter_ctx = ClassContext {
+        class_name: "duke/util/ArrayListIterator".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: vec![
+            FieldEntry {
+                name: "list".to_string(),
+                descriptor: "Ljava/util/ArrayList;".to_string(),
+                is_static: false,
+            },
+            FieldEntry {
+                name: "index".to_string(),
+                descriptor: "I".to_string(),
+                is_static: false,
+            },
+        ],
+        static_fields: Vec::new(),
+        instance_field_count: 2,
+        interfaces: Vec::new(),
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(iter_ctx);
+    registry.natives_mut().register(
+        "duke/util/ArrayListIterator",
+        "<init>",
+        "()V",
+        native_arraylist_iter_init,
+    );
+    registry.natives_mut().register(
+        "duke/util/ArrayListIterator",
+        "hasNext",
+        "()Z",
+        native_arraylist_iter_hasnext,
+    );
+    registry.natives_mut().register(
+        "duke/util/ArrayListIterator",
+        "next",
+        "()Ljava/lang/Object;",
+        native_arraylist_iter_next,
+    );
+
+    // java/util/Arrays — static array utilities
+    let arrays_ctx = ClassContext {
+        class_name: "java/util/Arrays".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: Vec::new(),
+        static_fields: Vec::new(),
+        instance_field_count: 0,
+        interfaces: Vec::new(),
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(arrays_ctx);
+    registry
+        .natives_mut()
+        .register("java/util/Arrays", "fill", "([II)V", native_arrays_fill_int);
+    registry.natives_mut().register(
+        "java/util/Arrays",
+        "fill",
+        "([Ljava/lang/Object;Ljava/lang/Object;)V",
+        native_arrays_fill_object,
+    );
+    registry.natives_mut().register(
+        "java/util/Arrays",
+        "copyOf",
+        "([II)[I",
+        native_arrays_copyof_int,
+    );
+    registry.natives_mut().register(
+        "java/util/Arrays",
+        "copyOf",
+        "([Ljava/lang/Object;I)[Ljava/lang/Object;",
+        native_arrays_copyof_object,
+    );
+    registry
+        .natives_mut()
+        .register("java/util/Arrays", "sort", "([I)V", native_arrays_sort_int);
+
+    // java/util/HashMap — hash map backed by flat key/value pair list in fields
+    // fields[0] = Int(size), fields[1]=key0, fields[2]=val0, fields[3]=key1, ...
+    let hashmap_ctx = ClassContext {
+        class_name: "java/util/HashMap".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: vec![FieldEntry {
+            name: "size".to_string(),
+            descriptor: "I".to_string(),
+            is_static: false,
+        }],
+        static_fields: Vec::new(),
+        instance_field_count: 1,
+        interfaces: Vec::new(),
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(hashmap_ctx);
+    registry
+        .natives_mut()
+        .register("java/util/HashMap", "<init>", "()V", native_hashmap_init);
+    registry.natives_mut().register(
+        "java/util/HashMap",
+        "put",
+        "(Ljava/lang/Object;Ljava/lang/Object;)Ljava/lang/Object;",
+        native_hashmap_put,
+    );
+    registry.natives_mut().register(
+        "java/util/HashMap",
+        "get",
+        "(Ljava/lang/Object;)Ljava/lang/Object;",
+        native_hashmap_get,
+    );
+    registry.natives_mut().register(
+        "java/util/HashMap",
+        "containsKey",
+        "(Ljava/lang/Object;)Z",
+        native_hashmap_contains_key,
+    );
+    registry
+        .natives_mut()
+        .register("java/util/HashMap", "size", "()I", native_hashmap_size);
+    registry.natives_mut().register(
+        "java/util/HashMap",
+        "remove",
+        "(Ljava/lang/Object;)Ljava/lang/Object;",
+        native_hashmap_remove,
+    );
+    registry.natives_mut().register(
+        "java/util/HashMap",
+        "isEmpty",
+        "()Z",
+        native_hashmap_is_empty,
+    );
+    registry.natives_mut().register(
+        "java/util/HashMap",
+        "getOrDefault",
+        "(Ljava/lang/Object;Ljava/lang/Object;)Ljava/lang/Object;",
+        native_hashmap_get_or_default,
+    );
+
+    // java/util/HashSet — set backed by unique elements in fields
+    // fields[0] = Int(size), fields[1..] = elements (unique, no duplicates)
+    let hashset_ctx = ClassContext {
+        class_name: "java/util/HashSet".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: vec![FieldEntry {
+            name: "size".to_string(),
+            descriptor: "I".to_string(),
+            is_static: false,
+        }],
+        static_fields: Vec::new(),
+        instance_field_count: 1,
+        interfaces: Vec::new(),
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(hashset_ctx);
+    registry
+        .natives_mut()
+        .register("java/util/HashSet", "<init>", "()V", native_hashset_init);
+    registry.natives_mut().register(
+        "java/util/HashSet",
+        "add",
+        "(Ljava/lang/Object;)Z",
+        native_hashset_add,
+    );
+    registry.natives_mut().register(
+        "java/util/HashSet",
+        "contains",
+        "(Ljava/lang/Object;)Z",
+        native_hashset_contains,
+    );
+    registry.natives_mut().register(
+        "java/util/HashSet",
+        "remove",
+        "(Ljava/lang/Object;)Z",
+        native_hashset_remove,
+    );
+    registry
+        .natives_mut()
+        .register("java/util/HashSet", "size", "()I", native_hashset_size);
+    registry.natives_mut().register(
+        "java/util/HashSet",
+        "isEmpty",
+        "()Z",
+        native_hashset_is_empty,
+    );
+
+    // java/util/Collections — static utility class
+    // Only sort(List) is supported; delegates to ArrayList.sort(null).
+    let collections_ctx = ClassContext {
+        class_name: "java/util/Collections".to_string(),
+        super_class: Some("java/lang/Object".to_string()),
+        constant_pool: Vec::new(),
+        methods: Vec::new(),
+        fields: Vec::new(),
+        static_fields: Vec::new(),
+        instance_field_count: 0,
+        interfaces: Vec::new(),
+        bootstrap_methods: Vec::new(),
+    };
+    registry.register(collections_ctx);
+    registry.natives_mut().register_callback(
+        "java/util/Collections",
+        "sort",
+        "(Ljava/util/List;)V",
+        native_collections_sort,
+    );
+}
+
+pub(crate) fn native_println_string(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let string_ref = match args.get(1) {
+        Some(Slot::Reference(Some(r))) => *r,
+        Some(Slot::Reference(None)) => {
+            writeln!(out, "null").ok();
+            return Ok(None);
+        }
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Reference",
+                got: "other",
+            });
+        }
+    };
+    let obj = heap.get(string_ref)?;
+    let text = obj.string_value.as_deref().unwrap_or("null");
+    writeln!(out, "{text}").ok();
+    Ok(None)
+}
+
+pub(crate) fn native_println_int(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let val = match args.get(1) {
+        Some(Slot::Int(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int",
+                got: "other",
+            });
+        }
+    };
+    writeln!(out, "{val}").ok();
+    Ok(None)
+}
+
+#[allow(clippy::unnecessary_wraps)] // must match NativeHandler signature
+pub(crate) fn native_println_void(
+    _args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    writeln!(out).ok();
+    Ok(None)
+}
+
+fn path_from_string_slot(
+    args: &[Slot],
+    idx: usize,
+    heap: &duke_gc::Heap,
+) -> VmResult<std::path::PathBuf> {
+    let path_ref = match args.get(idx) {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let path = heap
+        .get(path_ref)?
+        .string_value
+        .clone()
+        .ok_or(VmError::NullPointerException)?;
+    Ok(std::path::PathBuf::from(path))
+}
+
+fn file_path_from_this(args: &[Slot], heap: &duke_gc::Heap) -> VmResult<std::path::PathBuf> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let path_ref = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let path = heap
+        .get(path_ref)?
+        .string_value
+        .clone()
+        .ok_or(VmError::NullPointerException)?;
+    Ok(std::path::PathBuf::from(path))
+}
+
+pub(crate) fn native_file_init(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let path_slot = args.get(1).copied().unwrap_or(Slot::Reference(None));
+    let file_obj = heap.get_mut(this_ref)?;
+    let Some(path_field) = file_obj.fields.first_mut() else {
+        return Err(VmError::InvalidRef { address: this_ref });
+    };
+    *path_field = path_slot;
+    Ok(None)
+}
+
+pub(crate) fn native_file_exists(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let path = file_path_from_this(args, heap)?;
+    Ok(Some(Slot::Int(i32::from(path.exists()))))
+}
+
+pub(crate) fn native_file_is_file(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let path = file_path_from_this(args, heap)?;
+    Ok(Some(Slot::Int(i32::from(path.is_file()))))
+}
+
+pub(crate) fn native_file_is_directory(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let path = file_path_from_this(args, heap)?;
+    Ok(Some(Slot::Int(i32::from(path.is_dir()))))
+}
+
+fn file_stream_id_from_this(args: &[Slot], heap: &duke_gc::Heap) -> VmResult<i32> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(id)) if *id > 0 => Ok(*id),
+        _ => Err(VmError::JavaException {
+            class_name: "java/io/IOException".to_string(),
+        }),
+    }
+}
+
+pub(crate) fn native_file_input_stream_init(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let path = path_from_string_slot(args, 1, heap)?;
+    let file_id = heap.open_host_input_file(&path)?;
+    let stream_obj = heap.get_mut(this_ref)?;
+    let Some(fd_field) = stream_obj.fields.first_mut() else {
+        return Err(VmError::InvalidRef { address: this_ref });
+    };
+    *fd_field = Slot::Int(file_id);
+    Ok(None)
+}
+
+pub(crate) fn native_file_input_stream_read(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let file_id = file_stream_id_from_this(args, heap)?;
+    Ok(Some(Slot::Int(heap.read_host_file_byte(file_id)?)))
+}
+
+pub(crate) fn native_file_input_stream_read_bytes(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let file_id = file_stream_id_from_this(args, heap)?;
+    let array_ref = match args.get(1) {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let len = heap.get(array_ref)?.fields.len();
+    if len == 0 {
+        return Ok(Some(Slot::Int(0)));
+    }
+
+    let mut count = 0_usize;
+    for idx in 0..len {
+        let next = heap.read_host_file_byte(file_id)?;
+        if next < 0 {
+            break;
+        }
+        heap.get_mut(array_ref)?.fields[idx] = Slot::Int(next);
+        count += 1;
+    }
+
+    if count == 0 {
+        Ok(Some(Slot::Int(-1)))
+    } else {
+        Ok(Some(Slot::Int(i32::try_from(count).unwrap_or(i32::MAX))))
+    }
+}
+
+pub(crate) fn native_file_input_stream_close(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let file_id = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(id)) => *id,
+        _ => {
+            return Err(VmError::JavaException {
+                class_name: "java/io/IOException".to_string(),
+            });
+        }
+    };
+    heap.close_host_file(file_id);
+    let stream_obj = heap.get_mut(this_ref)?;
+    let Some(fd_field) = stream_obj.fields.first_mut() else {
+        return Err(VmError::InvalidRef { address: this_ref });
+    };
+    *fd_field = Slot::Int(0);
+    Ok(None)
+}
+
+pub(crate) fn native_file_output_stream_init(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let path = path_from_string_slot(args, 1, heap)?;
+    let file_id = heap.open_host_output_file(&path)?;
+    let stream_obj = heap.get_mut(this_ref)?;
+    let Some(fd_field) = stream_obj.fields.first_mut() else {
+        return Err(VmError::InvalidRef { address: this_ref });
+    };
+    *fd_field = Slot::Int(file_id);
+    Ok(None)
+}
+
+pub(crate) fn native_file_output_stream_write(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let file_id = file_stream_id_from_this(args, heap)?;
+    let value = match args.get(1) {
+        Some(Slot::Int(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int",
+                got: "other",
+            });
+        }
+    };
+    heap.write_host_file_byte(file_id, value)?;
+    Ok(None)
+}
+
+pub(crate) fn native_file_output_stream_write_bytes(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let file_id = file_stream_id_from_this(args, heap)?;
+    let array_ref = match args.get(1) {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let bytes = heap.get(array_ref)?.fields.clone();
+    for byte in bytes {
+        let Slot::Int(value) = byte else {
+            return Err(VmError::TypeMismatch {
+                expected: "Int",
+                got: "other",
+            });
+        };
+        heap.write_host_file_byte(file_id, value)?;
+    }
+    Ok(None)
+}
+
+pub(crate) fn native_file_output_stream_close(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let file_id = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(id)) => *id,
+        _ => {
+            return Err(VmError::JavaException {
+                class_name: "java/io/IOException".to_string(),
+            });
+        }
+    };
+    heap.close_host_file(file_id);
+    let stream_obj = heap.get_mut(this_ref)?;
+    let Some(fd_field) = stream_obj.fields.first_mut() else {
+        return Err(VmError::InvalidRef { address: this_ref });
+    };
+    *fd_field = Slot::Int(0);
+    Ok(None)
+}
+
+// ── Networking natives ────────────────────────────────────────────────────
+
+/// Native: `ServerSocket.<init>(int port)` — binds to 0.0.0.0:{port}.
+pub(crate) fn native_server_socket_init(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let port = match args.get(1) {
+        Some(Slot::Int(p)) => *p,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int",
+                got: "other",
+            });
+        }
+    };
+    let addr = format!("0.0.0.0:{port}");
+    let server_id = heap.bind_server_socket(&addr)?;
+    let actual_port = heap.server_socket_local_port(server_id)?;
+    let obj = heap.get_mut(this_ref)?;
+    if obj.fields.len() < 2 {
+        return Err(VmError::InvalidRef { address: this_ref });
+    }
+    obj.fields[0] = Slot::Int(server_id);
+    obj.fields[1] = Slot::Int(actual_port);
+    Ok(None)
+}
+
+/// Native: `ServerSocket.accept()` — blocks until a client connects, returns a Socket.
+pub(crate) fn native_server_socket_accept(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let server_fd = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(id)) if *id > 0 => *id,
+        _ => {
+            return Err(VmError::JavaException {
+                class_name: "java/io/IOException".into(),
+            });
+        }
+    };
+    let (reader_id, writer_id) = heap.accept_connection(server_fd)?;
+    // Allocate a new Socket object with fdRead=reader_id, fdWrite=writer_id
+    let socket_ref = heap.allocate("java/net/Socket".to_string(), 2);
+    heap.get_mut(socket_ref)?.fields[0] = Slot::Int(reader_id);
+    heap.get_mut(socket_ref)?.fields[1] = Slot::Int(writer_id);
+    Ok(Some(Slot::Reference(Some(socket_ref))))
+}
+
+/// Native: `ServerSocket.getLocalPort()` — returns the bound port.
+pub(crate) fn native_server_socket_get_local_port(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    match heap.get(this_ref)?.fields.get(1) {
+        Some(Slot::Int(port)) => Ok(Some(Slot::Int(*port))),
+        _ => Err(VmError::JavaException {
+            class_name: "java/io/IOException".into(),
+        }),
+    }
+}
+
+/// Native: `ServerSocket.close()` — closes the OS listener and zeros the fd field.
+pub(crate) fn native_server_socket_close(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let fd = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(id)) if *id > 0 => *id,
+        Some(Slot::Int(_)) => return Ok(None), // already closed — idempotent
+        _ => {
+            return Err(VmError::JavaException {
+                class_name: "java/io/IOException".into(),
+            });
+        }
+    };
+    heap.close_host_file(fd);
+    let obj = heap.get_mut(this_ref)?;
+    obj.fields[0] = Slot::Int(0);
+    obj.fields[1] = Slot::Int(0); // also zero cached port so getLocalPort() returns 0 after close
+    Ok(None)
+}
+
+/// Native: `Socket.<init>(String host, int port)` — connects to host:port.
+pub(crate) fn native_socket_init(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let host = match args.get(1) {
+        Some(Slot::Reference(Some(r))) => heap
+            .get(*r)?
+            .string_value
+            .clone()
+            .ok_or(VmError::NullPointerException)?,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let port = match args.get(2) {
+        Some(Slot::Int(p)) => *p,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int",
+                got: "other",
+            });
+        }
+    };
+    let addr = format!("{host}:{port}");
+    let (reader_id, writer_id) = heap.connect_socket(&addr)?;
+    let obj = heap.get_mut(this_ref)?;
+    if obj.fields.len() < 2 {
+        return Err(VmError::InvalidRef { address: this_ref });
+    }
+    obj.fields[0] = Slot::Int(reader_id);
+    obj.fields[1] = Slot::Int(writer_id);
+    Ok(None)
+}
+
+/// Native: `Socket.getInputStream()` — allocates a `SocketInputStream` wrapping `fdRead`.
+pub(crate) fn native_socket_get_input_stream(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let fd_read = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(id)) if *id > 0 => *id,
+        _ => {
+            return Err(VmError::JavaException {
+                class_name: "java/io/IOException".into(),
+            });
+        }
+    };
+    let stream_ref = heap.allocate("duke/net/SocketInputStream".to_string(), 1);
+    heap.get_mut(stream_ref)?.fields[0] = Slot::Int(fd_read);
+    Ok(Some(Slot::Reference(Some(stream_ref))))
+}
+
+/// Native: `Socket.getOutputStream()` — allocates a `SocketOutputStream` wrapping `fdWrite`.
+pub(crate) fn native_socket_get_output_stream(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let fd_write = match heap.get(this_ref)?.fields.get(1) {
+        Some(Slot::Int(id)) if *id > 0 => *id,
+        _ => {
+            return Err(VmError::JavaException {
+                class_name: "java/io/IOException".into(),
+            });
+        }
+    };
+    let stream_ref = heap.allocate("duke/net/SocketOutputStream".to_string(), 1);
+    heap.get_mut(stream_ref)?.fields[0] = Slot::Int(fd_write);
+    Ok(Some(Slot::Reference(Some(stream_ref))))
+}
+
+/// Native: `Socket.close()` — closes both OS handles (fdRead and fdWrite).
+pub(crate) fn native_socket_close(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let fd_read = match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(id)) => *id,
+        _ => {
+            return Err(VmError::JavaException {
+                class_name: "java/io/IOException".into(),
+            });
+        }
+    };
+    let fd_write = match heap.get(this_ref)?.fields.get(1) {
+        Some(Slot::Int(id)) => *id,
+        _ => {
+            return Err(VmError::JavaException {
+                class_name: "java/io/IOException".into(),
+            });
+        }
+    };
+    heap.close_host_file(fd_read);
+    heap.close_host_file(fd_write);
+    let obj = heap.get_mut(this_ref)?;
+    obj.fields[0] = Slot::Int(0);
+    obj.fields[1] = Slot::Int(0);
+    Ok(None)
+}
+
+/// Native: `String.length()` — returns string length as int.
+#[allow(clippy::cast_possible_truncation, clippy::cast_possible_wrap)]
+pub(crate) fn native_string_length(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let obj = heap.get(this_ref)?;
+    let len = obj.string_value.as_ref().map_or(0, String::len);
+    Ok(Some(Slot::Int(len as i32)))
+}
+
+/// Native: `String.equals(Object)` — compares string content.
+pub(crate) fn native_string_equals(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let other_ref = match args.get(1) {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Ok(Some(Slot::Int(0))),
+    };
+    // Fetch objects from heap in one go to keep borrows short
+    let this_obj = heap.get(this_ref)?;
+    let other_obj = heap.get(other_ref)?;
+    let this_str = this_obj.string_value.as_deref().unwrap_or_default();
+    let other_str = other_obj.string_value.as_deref().unwrap_or_default();
+    Ok(Some(Slot::Int(i32::from(this_str == other_str))))
+}
+
+/// Native: `String.charAt(int)` — returns char at index as int.
+#[allow(clippy::cast_sign_loss)]
+pub(crate) fn native_string_char_at(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let index = match args.get(1) {
+        Some(Slot::Int(i)) => *i,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int",
+                got: "other",
+            });
+        }
+    };
+    let obj = heap.get(this_ref)?;
+    let s = obj.string_value.as_deref().unwrap_or("");
+    let ch = s
+        .chars()
+        .nth(index as usize)
+        .ok_or(VmError::ArrayIndexOutOfBounds {
+            index,
+            length: s.len(),
+        })?;
+    Ok(Some(Slot::Int(ch as i32)))
+}
+
+/// Native: `Object.hashCode()` — returns heap address as hash.
+pub(crate) fn native_object_hashcode(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    match args.first() {
+        #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
+        Some(Slot::Reference(Some(r))) => Ok(Some(Slot::Int(*r as i32))),
+        _ => Err(VmError::NullPointerException),
+    }
+}
+
+/// Native: `Object.toString()` — delegates to `heap_object_to_string` so String,
+/// boxed primitives, and opaque objects all produce the correct Java representation.
+pub(crate) fn native_object_tostring(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let s = heap_object_to_string(heap.get(this_ref)?, this_ref);
+    let r = heap.allocate_string(s);
+    Ok(Some(Slot::Reference(Some(r))))
+}
+
+/// Native: `Object.clone()` — shallow-copies a heap object.
+pub(crate) fn native_object_clone(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let obj = heap.get(this_ref)?;
+    let cloned_class = obj.class_name.clone();
+    let cloned_fields = obj.fields.clone();
+    let cloned_string = obj.string_value.clone();
+    let new_ref = heap.allocate(cloned_class, 0);
+    let dest = heap.get_mut(new_ref)?;
+    dest.fields = cloned_fields;
+    dest.string_value = cloned_string;
+    Ok(Some(Slot::Reference(Some(new_ref))))
+}
+
+/// `Throwable.addSuppressed(Throwable suppressed)V`
+///
+/// No-op stub. Control flow is handled entirely by the bytecode desugaring —
+/// `addSuppressed` only affects what `getSuppressed()` returns, which is not
+/// yet implemented. Suppressed exception is silently dropped.
+///
+/// Signature: `args[0]` = this (Throwable), `args[1]` = suppressed (Throwable)
+#[allow(clippy::unnecessary_wraps)]
+pub(crate) fn native_throwable_add_suppressed(
+    _args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _stdout: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    Ok(None)
+}
+
+/// Native: `Enum.<init>(Ljava/lang/String;I)V` — stores name + ordinal.
+/// args: `[this_ref, name_ref, ordinal_int]`
+pub(crate) fn native_enum_init(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let name_slot = args.get(1).copied().unwrap_or(Slot::Reference(None));
+    let ordinal = match args.get(2) {
+        Some(Slot::Int(v)) => *v,
+        _ => 0,
+    };
+    let obj = heap.get_mut(this_ref)?;
+    if obj.fields.len() >= 2 {
+        obj.fields[0] = name_slot;
+        obj.fields[1] = Slot::Int(ordinal);
+    }
+    Ok(None)
+}
+
+/// Native: `Enum.ordinal()I`
+pub(crate) fn native_enum_ordinal(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let obj = heap.get(this_ref)?;
+    match obj.fields.get(1) {
+        Some(Slot::Int(v)) => Ok(Some(Slot::Int(*v))),
+        _ => Ok(Some(Slot::Int(0))),
+    }
+}
+
+/// Native: `Enum.name()Ljava/lang/String;`
+pub(crate) fn native_enum_name(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let obj = heap.get(this_ref)?;
+    match obj.fields.first() {
+        Some(slot @ Slot::Reference(_)) => Ok(Some(*slot)),
+        _ => Ok(Some(Slot::Reference(None))),
+    }
+}
+
+/// Native: `Enum.valueOf(Ljava/lang/Class;Ljava/lang/String;)Ljava/lang/Enum;`
+/// Searches heap for enum constants of the given class matching the name.
+#[allow(clippy::cast_possible_truncation)]
+pub(crate) fn native_enum_valueof(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let class_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let name_ref = match args.get(1) {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let target_name = heap.get(name_ref)?.string_value.clone().unwrap_or_default();
+    let enum_class_name = heap
+        .get(class_ref)?
+        .string_value
+        .clone()
+        .unwrap_or_default();
+
+    let obj_count = heap.len();
+    for i in 0..obj_count {
+        let obj = heap.get(i as u64)?;
+        if obj.class_name == enum_class_name
+            && obj.fields.len() >= 2
+            && let Some(Slot::Reference(Some(name_r))) = obj.fields.first()
+            && let Ok(name_obj) = heap.get(*name_r)
+            && name_obj.string_value.as_deref() == Some(target_name.as_str())
+        {
+            return Ok(Some(Slot::Reference(Some(i as u64))));
+        }
+    }
+
+    Err(VmError::JavaException {
+        class_name: "java/lang/IllegalArgumentException".to_string(),
+    })
+}
+
+/// Native: `String.valueOf(int)` — static method, returns string of int.
+pub(crate) fn native_string_value_of_int(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let val = match args.first() {
+        Some(Slot::Int(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int",
+                got: "other",
+            });
+        }
+    };
+    let s = val.to_string();
+    let r = heap.allocate_string(s);
+    Ok(Some(Slot::Reference(Some(r))))
+}
+
+/// Native: `PrintStream.print(String)` — no newline.
+pub(crate) fn native_print_string(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let string_ref = match args.get(1) {
+        Some(Slot::Reference(Some(r))) => *r,
+        Some(Slot::Reference(None)) => {
+            write!(out, "null").ok();
+            return Ok(None);
+        }
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Reference",
+                got: "other",
+            });
+        }
+    };
+    let obj = heap.get(string_ref)?;
+    let text = obj.string_value.as_deref().unwrap_or("null");
+    write!(out, "{text}").ok();
+    Ok(None)
+}
+
+/// Native: `PrintStream.print(int)` — no newline.
+#[allow(clippy::unnecessary_wraps)]
+pub(crate) fn native_print_int(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let val = match args.get(1) {
+        Some(Slot::Int(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int",
+                got: "other",
+            });
+        }
+    };
+    write!(out, "{val}").ok();
+    Ok(None)
+}
+
+// ---------------------------------------------------------------------------
+// println overloads (long, float, double, boolean, char, object)
+// ---------------------------------------------------------------------------
+
+pub(crate) fn native_println_long(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let val = match args.get(1) {
+        Some(Slot::Long(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Long",
+                got: "other",
+            });
+        }
+    };
+    writeln!(out, "{val}").ok();
+    Ok(None)
+}
+
+pub(crate) fn native_println_float(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let val = match args.get(1) {
+        Some(Slot::Float(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Float",
+                got: "other",
+            });
+        }
+    };
+    writeln!(out, "{val}").ok();
+    Ok(None)
+}
+
+pub(crate) fn native_println_double(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let val = match args.get(1) {
+        Some(Slot::Double(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Double",
+                got: "other",
+            });
+        }
+    };
+    writeln!(out, "{val}").ok();
+    Ok(None)
+}
+
+pub(crate) fn native_println_boolean(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let val = match args.get(1) {
+        Some(Slot::Int(v)) => *v != 0,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int(boolean)",
+                got: "other",
+            });
+        }
+    };
+    writeln!(out, "{val}").ok();
+    Ok(None)
+}
+
+pub(crate) fn native_println_char(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let val = match args.get(1) {
+        Some(Slot::Int(v)) => char::from_u32((*v).cast_unsigned()).unwrap_or('?'),
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int(char)",
+                got: "other",
+            });
+        }
+    };
+    writeln!(out, "{val}").ok();
+    Ok(None)
+}
+
+#[allow(clippy::cast_possible_wrap)]
+/// Convert a heap object to its Java display string.
+///
+/// Checks `string_value` first (handles String/StringBuilder).
+/// For boxed primitives, extracts the stored value from `fields[0]`.
+/// Falls back to `class_name@hex_ref` for opaque objects.
+pub(crate) fn heap_object_to_string(obj: &duke_gc::HeapObject, obj_ref: u64) -> String {
+    if let Some(s) = &obj.string_value {
+        return s.clone();
+    }
+    match obj.class_name.as_str() {
+        "java/lang/Integer" => {
+            if let Some(Slot::Int(v)) = obj.fields.first() {
+                return v.to_string();
+            }
+        }
+        "java/lang/Long" => {
+            if let Some(Slot::Long(v)) = obj.fields.first() {
+                return v.to_string();
+            }
+        }
+        "java/lang/Double" => {
+            if let Some(Slot::Double(v)) = obj.fields.first() {
+                return format_java_double(*v);
+            }
+        }
+        "java/lang/Float" => {
+            if let Some(Slot::Float(v)) = obj.fields.first() {
+                return format_java_float(*v);
+            }
+        }
+        "java/lang/Boolean" => {
+            return match obj.fields.first() {
+                Some(Slot::Int(v)) if *v != 0 => "true".to_string(),
+                _ => "false".to_string(),
+            };
+        }
+        "java/lang/Character" => {
+            if let Some(Slot::Int(v)) = obj.fields.first()
+                && let Some(c) = char::from_u32((*v).cast_unsigned())
+            {
+                return c.to_string();
+            }
+        }
+        _ => {}
+    }
+    format!("{}@{:x}", obj.class_name, obj_ref)
+}
+
+pub(crate) fn native_println_object(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    match args.get(1) {
+        Some(Slot::Reference(Some(r))) => {
+            let s = heap_object_to_string(heap.get(*r)?, *r);
+            writeln!(out, "{s}").ok();
+        }
+        Some(Slot::Reference(None)) => {
+            writeln!(out, "null").ok();
+        }
+        _ => {
+            writeln!(out, "<unknown>").ok();
+        }
+    }
+    Ok(None)
+}
+
+// ---------------------------------------------------------------------------
+// print overloads (long, float, double, boolean, char, object)
+// ---------------------------------------------------------------------------
+
+pub(crate) fn native_print_long(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let val = match args.get(1) {
+        Some(Slot::Long(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Long",
+                got: "other",
+            });
+        }
+    };
+    write!(out, "{val}").ok();
+    Ok(None)
+}
+
+pub(crate) fn native_print_float(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let val = match args.get(1) {
+        Some(Slot::Float(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Float",
+                got: "other",
+            });
+        }
+    };
+    write!(out, "{val}").ok();
+    Ok(None)
+}
+
+pub(crate) fn native_print_double(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let val = match args.get(1) {
+        Some(Slot::Double(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Double",
+                got: "other",
+            });
+        }
+    };
+    write!(out, "{val}").ok();
+    Ok(None)
+}
+
+pub(crate) fn native_print_boolean(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let val = match args.get(1) {
+        Some(Slot::Int(v)) => *v != 0,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int(boolean)",
+                got: "other",
+            });
+        }
+    };
+    write!(out, "{val}").ok();
+    Ok(None)
+}
+
+pub(crate) fn native_print_char(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let val = match args.get(1) {
+        Some(Slot::Int(v)) => char::from_u32((*v).cast_unsigned()).unwrap_or('?'),
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int(char)",
+                got: "other",
+            });
+        }
+    };
+    write!(out, "{val}").ok();
+    Ok(None)
+}
+
+#[allow(clippy::cast_possible_truncation, clippy::cast_possible_wrap)]
+pub(crate) fn native_print_object(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    match args.get(1) {
+        Some(Slot::Reference(Some(r))) => {
+            let s = heap_object_to_string(heap.get(*r)?, *r);
+            write!(out, "{s}").ok();
+        }
+        Some(Slot::Reference(None)) => {
+            write!(out, "null").ok();
+        }
+        _ => {
+            write!(out, "<unknown>").ok();
+        }
+    }
+    Ok(None)
+}
+
+// ---------------------------------------------------------------------------
+// System.exit
+// ---------------------------------------------------------------------------
+
+pub(crate) fn native_system_exit(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let code = match args.first() {
+        Some(Slot::Int(v)) => *v,
+        _ => 1,
+    };
+    Err(VmError::SystemExit { code })
+}
+
+pub(crate) const THREAD_TARGET_SLOT: usize = 0;
+pub(crate) const THREAD_ID_SLOT: usize = 1;
+
+pub(crate) fn native_thread_init(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let this = heap.get_mut(this_ref)?;
+    this.fields[THREAD_TARGET_SLOT] = Slot::Reference(None);
+    this.fields[THREAD_ID_SLOT] = Slot::Int(-1);
+    Ok(None)
+}
+
+pub(crate) fn native_thread_init_runnable(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let target = args.get(1).copied().unwrap_or(Slot::Reference(None));
+    let this = heap.get_mut(this_ref)?;
+    this.fields[THREAD_TARGET_SLOT] = target;
+    this.fields[THREAD_ID_SLOT] = Slot::Int(-1);
+    Ok(None)
+}
+
+pub(crate) fn native_thread_start(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let thread_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    control.request(NativeThreadAction::Start { thread_ref });
+    Ok(None)
+}
+
+pub(crate) fn native_thread_join(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let thread_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let thread = heap.get(thread_ref)?;
+    let Some(Slot::Int(thread_id)) = thread.fields.get(THREAD_ID_SLOT) else {
+        return Ok(None);
+    };
+    if *thread_id >= 0 {
+        control.request(NativeThreadAction::Join {
+            thread_id: *thread_id,
+        });
+    }
+    Ok(None)
+}
+
+pub(crate) fn native_thread_sleep(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let millis = match args.first() {
+        Some(Slot::Long(value)) => *value,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "long",
+                got: "other",
+            });
+        }
+    };
+    let millis = u64::try_from(millis.max(0)).unwrap_or(0);
+    control.request(NativeThreadAction::Sleep(std::time::Duration::from_millis(
+        millis,
+    )));
+    Ok(None)
+}
+
+/// Native: `String.substring(int)` — substring from begin to end.
+#[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
+pub(crate) fn native_string_substring(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+
+    let begin = match args.get(1) {
+        Some(Slot::Int(v)) => *v as usize,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int",
+                got: "other",
+            });
+        }
+    };
+    let sub = {
+        let obj = heap.get(this_ref)?;
+        let s = obj.string_value.as_deref().unwrap_or_default();
+        if begin > s.len() {
+            return Err(VmError::ArrayIndexOutOfBounds {
+                index: i32::try_from(begin).unwrap_or(i32::MAX),
+                length: s.len(),
+            });
+        }
+        s.chars().skip(begin).collect::<String>()
+    };
+
+    let r = heap.allocate_string(sub);
+    Ok(Some(Slot::Reference(Some(r))))
+}
+
+/// Native: `String.substring(int, int)` — substring from begin to end (exclusive).
+#[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
+pub(crate) fn native_string_substring_range(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+
+    let begin = match args.get(1) {
+        Some(Slot::Int(v)) => *v as usize,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int",
+                got: "other",
+            });
+        }
+    };
+    let end = match args.get(2) {
+        Some(Slot::Int(v)) => *v as usize,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int",
+                got: "other",
+            });
+        }
+    };
+    let sub = {
+        let obj = heap.get(this_ref)?;
+        let s = obj.string_value.as_deref().unwrap_or_default();
+        if begin > end || end > s.len() {
+            return Err(VmError::ArrayIndexOutOfBounds {
+                index: i32::try_from(end).unwrap_or(i32::MAX),
+                length: s.len(),
+            });
+        }
+        s.chars().skip(begin).take(end - begin).collect::<String>()
+    };
+
+    let r = heap.allocate_string(sub);
+    Ok(Some(Slot::Reference(Some(r))))
+}
+
+/// Native: `String.indexOf(String)` — find first occurrence of target.
+pub(crate) fn native_string_indexof(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+
+    let target_ref = match args.get(1) {
+        Some(Slot::Reference(Some(r))) => *r,
+        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Reference",
+                got: "other",
+            });
+        }
+    };
+    // Fetch objects from heap in one go to keep borrows short
+    let this_obj = heap.get(this_ref)?;
+    let target_obj = heap.get(target_ref)?;
+    let s = this_obj.string_value.as_deref().unwrap_or_default();
+    let target = target_obj.string_value.as_deref().unwrap_or_default();
+
+    #[allow(clippy::cast_possible_truncation, clippy::cast_possible_wrap)]
+    let result = s.find(target).map_or(-1, |i| i as i32);
+    Ok(Some(Slot::Int(result)))
+}
+
+/// Native: `String.contains(CharSequence)` — check if string contains target.
+pub(crate) fn native_string_contains(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+
+    let target_ref = match args.get(1) {
+        Some(Slot::Reference(Some(r))) => *r,
+        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Reference",
+                got: "other",
+            });
+        }
+    };
+    // Fetch objects from heap in one go to keep borrows short
+    let this_obj = heap.get(this_ref)?;
+    let target_obj = heap.get(target_ref)?;
+    let s = this_obj.string_value.as_deref().unwrap_or_default();
+    let target = target_obj.string_value.as_deref().unwrap_or_default();
+    Ok(Some(Slot::Int(i32::from(s.contains(target)))))
+}
+
+/// Native: `String.isEmpty()` — check if string is empty.
+#[allow(clippy::unnecessary_wraps)]
+pub(crate) fn native_string_isempty(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let obj = heap.get(this_ref)?;
+    let s = obj.string_value.as_deref().unwrap_or_default();
+    Ok(Some(Slot::Int(i32::from(s.is_empty()))))
+}
+
+// Dispatch string constants used by the callback-based sort chain.
+const COMPARE_TO_METHOD: &str = "compareTo";
+const COMPARE_TO_OBJECT_DESC: &str = "(Ljava/lang/Object;)I";
+const SORT_COMPARATOR_DESC: &str = "(Ljava/util/Comparator;)V";
+
+/// Maps a `std::cmp::Ordering` to the Java `compareTo` convention: -1 / 0 / 1.
+///
+/// Used by all boxed-type `compareTo` natives to return a consistent,
+/// sign-correct value without relying on `Ordering`'s internal discriminant.
+#[inline]
+const fn ordering_to_int(o: std::cmp::Ordering) -> i32 {
+    match o {
+        std::cmp::Ordering::Less => -1,
+        std::cmp::Ordering::Equal => 0,
+        std::cmp::Ordering::Greater => 1,
+    }
+}
+
+/// Native: `String.compareTo(String)` — delegates to the Object overload.
+pub(crate) fn native_string_compareto(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    out: &mut dyn Write,
+    control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    native_string_compareto_object(args, heap, out, control)
+}
+
+/// Native: `String.compareTo(Object)` — lexicographic comparison via Object descriptor.
+pub(crate) fn native_string_compareto_object(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let str_val = |s: &Slot| -> VmResult<String> {
+        match s {
+            Slot::Reference(Some(r)) => Ok(heap.get(*r)?.string_value.clone().unwrap_or_default()),
+            _ => Err(VmError::NullPointerException),
+        }
+    };
+    let a = match args.first() {
+        Some(s) => str_val(s)?,
+        None => return Err(VmError::NullPointerException),
+    };
+    let b = match args.get(1) {
+        Some(s) => str_val(s)?,
+        None => return Err(VmError::NullPointerException),
+    };
+    Ok(Some(Slot::Int(ordering_to_int(a.as_str().cmp(b.as_str())))))
+}
+
+/// Native: `String.startsWith(String)` — check if string starts with prefix.
+pub(crate) fn native_string_startswith(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
+    let prefix_ref = match args.get(1) {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let prefix = heap
+        .get(prefix_ref)?
+        .string_value
+        .clone()
+        .unwrap_or_default();
+    Ok(Some(Slot::Int(i32::from(s.starts_with(&prefix)))))
+}
+
+/// Native: `String.endsWith(String)` — check if string ends with suffix.
+pub(crate) fn native_string_endswith(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
+    let suffix_ref = match args.get(1) {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let suffix = heap
+        .get(suffix_ref)?
+        .string_value
+        .clone()
+        .unwrap_or_default();
+    Ok(Some(Slot::Int(i32::from(s.ends_with(&suffix)))))
+}
+
+/// Native: `String.trim()` — remove leading and trailing whitespace.
+pub(crate) fn native_string_trim(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
+    let trimmed = s.trim().to_string();
+    let r = heap.allocate_string(trimmed);
+    Ok(Some(Slot::Reference(Some(r))))
+}
+
+/// Native: `String.toCharArray()` — convert string to char array.
+#[allow(clippy::cast_possible_truncation, clippy::cast_possible_wrap)]
+pub(crate) fn native_string_tochararray(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
+    let chars: Vec<char> = s.chars().collect();
+    let arr_ref = heap.allocate("[C".to_string(), chars.len());
+    for (i, &c) in chars.iter().enumerate() {
+        heap.get_mut(arr_ref).unwrap().fields[i] = Slot::Int(c as i32);
+    }
+    Ok(Some(Slot::Reference(Some(arr_ref))))
+}
+
+// ---- Integer natives ----
+
+/// Native: `Integer.parseInt(String)` — parses string to int.
+pub(crate) fn native_integer_parseint(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let str_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Reference",
+                got: "other",
+            });
+        }
+    };
+    let s = heap.get(str_ref)?.string_value.clone().unwrap_or_default();
+    let val: i32 = s.trim().parse().map_err(|_| VmError::JavaException {
+        class_name: "java/lang/NumberFormatException".to_string(),
+    })?;
+    Ok(Some(Slot::Int(val)))
+}
+
+/// Native: `Integer.valueOf(int)` — boxes int into Integer object.
+pub(crate) fn native_integer_valueof(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let val = match args.first() {
+        Some(Slot::Int(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int",
+                got: "other",
+            });
+        }
+    };
+    let r = heap.allocate("java/lang/Integer".to_string(), 1);
+    heap.get_mut(r).unwrap().fields[0] = Slot::Int(val);
+    Ok(Some(Slot::Reference(Some(r))))
+}
+
+/// Native: `Integer.intValue()` — unboxes Integer to int.
+pub(crate) fn native_integer_intvalue(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let val = heap.get(this_ref)?.fields[0];
+    Ok(Some(val))
+}
+
+/// Native: `Integer.toString(int)` — static, converts int to String.
+pub(crate) fn native_integer_tostring_static(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let val = match args.first() {
+        Some(Slot::Int(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int",
+                got: "other",
+            });
+        }
+    };
+    let r = heap.allocate_string(val.to_string());
+    Ok(Some(Slot::Reference(Some(r))))
+}
+
+/// Native: `Integer.compareTo(Object)` — compares two boxed Integers.
+pub(crate) fn native_integer_compareto(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let int_val = |s: &Slot| -> VmResult<i32> {
+        match s {
+            Slot::Reference(Some(r)) => match heap.get(*r)?.fields.first() {
+                Some(Slot::Int(n)) => Ok(*n),
+                _ => Err(VmError::InvalidRef { address: *r }),
+            },
+            _ => Err(VmError::NullPointerException),
+        }
+    };
+    let a = match args.first() {
+        Some(s) => int_val(s)?,
+        None => return Err(VmError::NullPointerException),
+    };
+    let b = match args.get(1) {
+        Some(s) => int_val(s)?,
+        None => return Err(VmError::NullPointerException),
+    };
+    Ok(Some(Slot::Int(ordering_to_int(a.cmp(&b)))))
+}
+
+// ---- String.valueOf overloads ----
+
+/// Native: `String.valueOf(long)` — converts long to String.
+pub(crate) fn native_string_value_of_long(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let val = match args.first() {
+        Some(Slot::Long(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Long",
+                got: "other",
+            });
+        }
+    };
+    let r = heap.allocate_string(val.to_string());
+    Ok(Some(Slot::Reference(Some(r))))
+}
+
+/// Native: `String.valueOf(double)` — converts double to String.
+pub(crate) fn native_string_value_of_double(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let val = match args.first() {
+        Some(Slot::Double(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Double",
+                got: "other",
+            });
+        }
+    };
+    let r = heap.allocate_string(val.to_string());
+    Ok(Some(Slot::Reference(Some(r))))
+}
+
+/// Native: `String.valueOf(float)` — converts float to String.
+pub(crate) fn native_string_value_of_float(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let val = match args.first() {
+        Some(Slot::Float(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Float",
+                got: "other",
+            });
+        }
+    };
+    let r = heap.allocate_string(val.to_string());
+    Ok(Some(Slot::Reference(Some(r))))
+}
+
+/// Native: `String.valueOf(boolean)` — converts boolean to String.
+pub(crate) fn native_string_value_of_boolean(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let val = match args.first() {
+        Some(Slot::Int(v)) => *v != 0,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int(boolean)",
+                got: "other",
+            });
+        }
+    };
+    let r = heap.allocate_string(if val { "true" } else { "false" }.to_string());
+    Ok(Some(Slot::Reference(Some(r))))
+}
+
+/// Native: `String.valueOf(char)` — converts char to String.
+pub(crate) fn native_string_value_of_char(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let val = match args.first() {
+        Some(Slot::Int(v)) => char::from_u32((*v).cast_unsigned()).unwrap_or('?'),
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int(char)",
+                got: "other",
+            });
+        }
+    };
+    let r = heap.allocate_string(val.to_string());
+    Ok(Some(Slot::Reference(Some(r))))
+}
+
+/// Native: `String.valueOf(Object)` — converts Object to String.
+pub(crate) fn native_string_value_of_object(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    match args.first() {
+        Some(Slot::Reference(Some(r))) => {
+            let s = heap_object_to_string(heap.get(*r)?, *r);
+            let r = heap.allocate_string(s);
+            Ok(Some(Slot::Reference(Some(r))))
+        }
+        Some(Slot::Reference(None)) => {
+            let r = heap.allocate_string("null".to_string());
+            Ok(Some(Slot::Reference(Some(r))))
+        }
+        _ => Err(VmError::TypeMismatch {
+            expected: "Reference",
+            got: "other",
+        }),
+    }
+}
+
+// ---- String.concat ----
+
+/// Native: `String.concat(String)` — concatenates two strings.
+pub(crate) fn native_string_concat(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let s1 = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
+    let other_ref = match args.get(1) {
+        Some(Slot::Reference(Some(r))) => *r,
+        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Reference",
+                got: "other",
+            });
+        }
+    };
+    let s2 = heap
+        .get(other_ref)?
+        .string_value
+        .clone()
+        .unwrap_or_default();
+    let r = heap.allocate_string(format!("{s1}{s2}"));
+    Ok(Some(Slot::Reference(Some(r))))
+}
+
+/// Formats a single boxed slot value using the given format specifier.
+pub(crate) fn format_arg(
+    spec: char,
+    precision: Option<usize>,
+    slot: &Slot,
+    heap: &duke_gc::Heap,
+) -> VmResult<String> {
+    match slot {
+        Slot::Reference(None) => Ok("null".to_string()),
+        Slot::Reference(Some(r)) => {
+            let obj = heap.get(*r)?;
+            match spec {
+                's' => Ok(heap_object_to_string(obj, *r)),
+                'd' => match obj.fields.first() {
+                    Some(Slot::Int(v)) => Ok(v.to_string()),
+                    Some(Slot::Long(v)) => Ok(v.to_string()),
+                    _ => Ok("0".to_string()),
+                },
+                'f' => {
+                    let v = match obj.fields.first() {
+                        Some(Slot::Double(v)) => *v,
+                        Some(Slot::Float(v)) => f64::from(*v),
+                        _ => 0.0,
+                    };
+                    Ok(precision.map_or_else(|| format!("{v:.6}"), |p| format!("{v:.p$}")))
+                }
+                'x' => match obj.fields.first() {
+                    Some(Slot::Int(v)) => Ok(format!("{v:x}")),
+                    Some(Slot::Long(v)) => Ok(format!("{v:x}")),
+                    _ => Ok("0".to_string()),
+                },
+                'X' => match obj.fields.first() {
+                    Some(Slot::Int(v)) => Ok(format!("{v:X}")),
+                    Some(Slot::Long(v)) => Ok(format!("{v:X}")),
+                    _ => Ok("0".to_string()),
+                },
+                _ => Ok(String::new()),
+            }
+        }
+        _ => Ok(String::new()),
+    }
+}
+
+/// Native: `String.format(Ljava/lang/String;[Ljava/lang/Object;)Ljava/lang/String;`
+pub(crate) fn native_string_format(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let fmt_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let fmt = heap.get(fmt_ref)?.string_value.clone().unwrap_or_default();
+
+    let arr_len = match args.get(1) {
+        Some(Slot::Reference(Some(r))) => heap.get(*r)?.fields.len(),
+        _ => 0,
+    };
+
+    let mut result = String::new();
+    let mut arg_idx = 0usize;
+    let chars: Vec<char> = fmt.chars().collect();
+    let mut i = 0;
+
+    while i < chars.len() {
+        if chars[i] != '%' {
+            result.push(chars[i]);
+            i += 1;
+            continue;
+        }
+        i += 1;
+        if i >= chars.len() {
+            break;
+        }
+
+        // Parse optional precision: %.2f
+        let precision: Option<usize> = if chars[i] == '.' {
+            i += 1;
+            let mut prec_str = String::new();
+            while i < chars.len() && chars[i].is_ascii_digit() {
+                prec_str.push(chars[i]);
+                i += 1;
+            }
+            prec_str.parse().ok()
+        } else {
+            None
+        };
+
+        // Skip optional width digits
+        while i < chars.len() && chars[i].is_ascii_digit() {
+            i += 1;
+        }
+
+        if i >= chars.len() {
+            break;
+        }
+        let spec = chars[i];
+        i += 1;
+
+        match spec {
+            '%' => result.push('%'),
+            'n' => result.push('\n'),
+            's' | 'd' | 'f' | 'x' | 'X' => {
+                let slot = if arg_idx < arr_len {
+                    match args.get(1) {
+                        Some(Slot::Reference(Some(r))) => heap
+                            .get(*r)?
+                            .fields
+                            .get(arg_idx)
+                            .copied()
+                            .unwrap_or(Slot::Reference(None)),
+                        _ => Slot::Reference(None),
+                    }
+                } else {
+                    Slot::Reference(None)
+                };
+                arg_idx += 1;
+                let formatted = format_arg(spec, precision, &slot, heap)?;
+                result.push_str(&formatted);
+            }
+            _ => {
+                result.push('%');
+                result.push(spec);
+            }
+        }
+    }
+
+    let r = heap.allocate_string(result);
+    Ok(Some(Slot::Reference(Some(r))))
+}
+
+// ---- Extended String natives ----
+
+/// Native: `String.toUpperCase()` — returns a new uppercase String.
+pub(crate) fn native_string_touppercase(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
+    let r = heap.allocate_string(s.to_uppercase());
+    Ok(Some(Slot::Reference(Some(r))))
+}
+
+/// Native: `String.toLowerCase()` — returns a new lowercase String.
+pub(crate) fn native_string_tolowercase(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
+    let r = heap.allocate_string(s.to_lowercase());
+    Ok(Some(Slot::Reference(Some(r))))
+}
+
+/// Native: `String.replace(char, char)` — replaces all occurrences of old char with new char.
+#[allow(clippy::cast_sign_loss)]
+pub(crate) fn native_string_replace_char(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
+    let old_char = match args.get(1) {
+        Some(Slot::Int(v)) => char::from_u32((*v).cast_unsigned()).unwrap_or('?'),
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int",
+                got: "other",
+            });
+        }
+    };
+    let new_char = match args.get(2) {
+        Some(Slot::Int(v)) => char::from_u32((*v).cast_unsigned()).unwrap_or('?'),
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int",
+                got: "other",
+            });
+        }
+    };
+    let result = s.replace(old_char, &new_char.to_string());
+    let r = heap.allocate_string(result);
+    Ok(Some(Slot::Reference(Some(r))))
+}
+
+/// Native: `String.replace(CharSequence, CharSequence)` — replaces all occurrences of target with replacement.
+pub(crate) fn native_string_replace_charsequence(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
+    let target_ref = match args.get(1) {
+        Some(Slot::Reference(Some(r))) => *r,
+        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Reference",
+                got: "other",
+            });
+        }
+    };
+    let target = heap
+        .get(target_ref)?
+        .string_value
+        .clone()
+        .unwrap_or_default();
+    let replacement_ref = match args.get(2) {
+        Some(Slot::Reference(Some(r))) => *r,
+        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Reference",
+                got: "other",
+            });
+        }
+    };
+    let replacement = heap
+        .get(replacement_ref)?
+        .string_value
+        .clone()
+        .unwrap_or_default();
+    let result = s.replace(&*target, &replacement);
+    let r = heap.allocate_string(result);
+    Ok(Some(Slot::Reference(Some(r))))
+}
+
+/// Native: `String.split(String)` — splits string by delimiter, returns String array.
+pub(crate) fn native_string_split(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
+    let delim_ref = match args.get(1) {
+        Some(Slot::Reference(Some(r))) => *r,
+        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Reference",
+                got: "other",
+            });
+        }
+    };
+    let delim = heap
+        .get(delim_ref)?
+        .string_value
+        .clone()
+        .unwrap_or_default();
+    let parts: Vec<&str> = s.split(&*delim).collect();
+    let arr_ref = heap.allocate("[Ljava/lang/String;".to_string(), parts.len());
+    for (i, part) in parts.iter().enumerate() {
+        let str_ref = heap.allocate_string((*part).to_string());
+        heap.get_mut(arr_ref).unwrap().fields[i] = Slot::Reference(Some(str_ref));
+    }
+    Ok(Some(Slot::Reference(Some(arr_ref))))
+}
+
+/// Native: `String.hashCode()` — Java's hash algorithm: `s[0]*31^(n-1) + s[1]*31^(n-2) + ... + s[n-1]`.
+#[allow(clippy::cast_possible_wrap)]
+pub(crate) fn native_string_hashcode(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let s = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
+    let mut h: i32 = 0;
+    for ch in s.chars() {
+        h = h.wrapping_mul(31).wrapping_add(ch as i32);
+    }
+    Ok(Some(Slot::Int(h)))
+}
+
+/// Native: `String.toString()` — identity, returns `this`.
+#[allow(clippy::unnecessary_wraps)]
+pub(crate) fn native_string_tostring(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    Ok(Some(args.first().copied().unwrap_or(Slot::Reference(None))))
+}
+
+// ---- Math natives ----
+
+/// Native: `Math.max(int, int)` — returns the larger value.
+pub(crate) fn native_math_max_int(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let a = match args.first() {
+        Some(Slot::Int(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int",
+                got: "other",
+            });
+        }
+    };
+    let b = match args.get(1) {
+        Some(Slot::Int(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int",
+                got: "other",
+            });
+        }
+    };
+    Ok(Some(Slot::Int(a.max(b))))
+}
+
+/// Native: `Math.min(int, int)` — returns the smaller value.
+pub(crate) fn native_math_min_int(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let a = match args.first() {
+        Some(Slot::Int(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int",
+                got: "other",
+            });
+        }
+    };
+    let b = match args.get(1) {
+        Some(Slot::Int(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int",
+                got: "other",
+            });
+        }
+    };
+    Ok(Some(Slot::Int(a.min(b))))
+}
+
+/// Native: `Math.abs(int)` — returns absolute value.
+pub(crate) fn native_math_abs_int(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let a = match args.first() {
+        Some(Slot::Int(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int",
+                got: "other",
+            });
+        }
+    };
+    Ok(Some(Slot::Int(a.wrapping_abs())))
+}
+
+// ---- Extended Math natives ----
+
+/// Native: `Math.sqrt(double)` — returns square root.
+pub(crate) fn native_math_sqrt(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let a = match args.first() {
+        Some(Slot::Double(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Double",
+                got: "other",
+            });
+        }
+    };
+    Ok(Some(Slot::Double(a.sqrt())))
+}
+
+/// Native: `Math.pow(double, double)` — returns a raised to the power b.
+pub(crate) fn native_math_pow(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let a = match args.first() {
+        Some(Slot::Double(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Double",
+                got: "other",
+            });
+        }
+    };
+    let b = match args.get(1) {
+        Some(Slot::Double(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Double",
+                got: "other",
+            });
+        }
+    };
+    Ok(Some(Slot::Double(a.powf(b))))
+}
+
+/// Native: `Math.floor(double)` — returns floor value.
+pub(crate) fn native_math_floor(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let a = match args.first() {
+        Some(Slot::Double(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Double",
+                got: "other",
+            });
+        }
+    };
+    Ok(Some(Slot::Double(a.floor())))
+}
+
+/// Native: `Math.ceil(double)` — returns ceiling value.
+pub(crate) fn native_math_ceil(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let a = match args.first() {
+        Some(Slot::Double(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Double",
+                got: "other",
+            });
+        }
+    };
+    Ok(Some(Slot::Double(a.ceil())))
+}
+
+/// Native: `Math.round(double)` — returns closest long.
+#[allow(clippy::cast_possible_truncation)]
+pub(crate) fn native_math_round_double(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let a = match args.first() {
+        Some(Slot::Double(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Double",
+                got: "other",
+            });
+        }
+    };
+    Ok(Some(Slot::Long(a.round() as i64)))
+}
+
+/// Native: `Math.abs(long)` — returns absolute value.
+pub(crate) fn native_math_abs_long(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let a = match args.first() {
+        Some(Slot::Long(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Long",
+                got: "other",
+            });
+        }
+    };
+    Ok(Some(Slot::Long(a.wrapping_abs())))
+}
+
+/// Native: `Math.abs(double)` — returns absolute value.
+pub(crate) fn native_math_abs_double(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let a = match args.first() {
+        Some(Slot::Double(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Double",
+                got: "other",
+            });
+        }
+    };
+    Ok(Some(Slot::Double(a.abs())))
+}
+
+/// Native: `Math.max(long, long)` — returns the larger value.
+pub(crate) fn native_math_max_long(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let a = match args.first() {
+        Some(Slot::Long(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Long",
+                got: "other",
+            });
+        }
+    };
+    let b = match args.get(1) {
+        Some(Slot::Long(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Long",
+                got: "other",
+            });
+        }
+    };
+    Ok(Some(Slot::Long(a.max(b))))
+}
+
+/// Native: `Math.min(long, long)` — returns the smaller value.
+pub(crate) fn native_math_min_long(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let a = match args.first() {
+        Some(Slot::Long(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Long",
+                got: "other",
+            });
+        }
+    };
+    let b = match args.get(1) {
+        Some(Slot::Long(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Long",
+                got: "other",
+            });
+        }
+    };
+    Ok(Some(Slot::Long(a.min(b))))
+}
+
+/// Native: `Math.max(double, double)` — returns the larger value.
+pub(crate) fn native_math_max_double(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let a = match args.first() {
+        Some(Slot::Double(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Double",
+                got: "other",
+            });
+        }
+    };
+    let b = match args.get(1) {
+        Some(Slot::Double(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Double",
+                got: "other",
+            });
+        }
+    };
+    Ok(Some(Slot::Double(a.max(b))))
+}
+
+/// Native: `Math.min(double, double)` — returns the smaller value.
+pub(crate) fn native_math_min_double(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let a = match args.first() {
+        Some(Slot::Double(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Double",
+                got: "other",
+            });
+        }
+    };
+    let b = match args.get(1) {
+        Some(Slot::Double(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Double",
+                got: "other",
+            });
+        }
+    };
+    Ok(Some(Slot::Double(a.min(b))))
+}
+
+// ---- Long class natives ----
+
+/// Native: `Long.parseLong(String)` — parses string to long.
+pub(crate) fn native_long_parselong(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let str_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Reference",
+                got: "other",
+            });
+        }
+    };
+    let s = heap.get(str_ref)?.string_value.clone().unwrap_or_default();
+    let val: i64 = s.trim().parse().map_err(|_| VmError::JavaException {
+        class_name: "java/lang/NumberFormatException".to_string(),
+    })?;
+    Ok(Some(Slot::Long(val)))
+}
+
+/// Native: `Long.valueOf(long)` — boxes long into Long object.
+pub(crate) fn native_long_valueof(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let val = match args.first() {
+        Some(Slot::Long(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Long",
+                got: "other",
+            });
+        }
+    };
+    let r = heap.allocate("java/lang/Long".to_string(), 1);
+    heap.get_mut(r).unwrap().fields[0] = Slot::Long(val);
+    Ok(Some(Slot::Reference(Some(r))))
+}
+
+/// Native: `Long.longValue()` — unboxes Long to long.
+pub(crate) fn native_long_longvalue(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let val = heap.get(this_ref)?.fields[0];
+    Ok(Some(val))
+}
+
+/// Native: `Long.toString(long)` — static, converts long to String.
+pub(crate) fn native_long_tostring_static(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let val = match args.first() {
+        Some(Slot::Long(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Long",
+                got: "other",
+            });
+        }
+    };
+    let r = heap.allocate_string(val.to_string());
+    Ok(Some(Slot::Reference(Some(r))))
+}
+
+/// Native: `Long.compareTo(Object)` — compares two boxed Longs.
+pub(crate) fn native_long_compareto(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let long_val = |s: &Slot| -> VmResult<i64> {
+        match s {
+            Slot::Reference(Some(r)) => match heap.get(*r)?.fields.first() {
+                Some(Slot::Long(n)) => Ok(*n),
+                _ => Err(VmError::InvalidRef { address: *r }),
+            },
+            _ => Err(VmError::NullPointerException),
+        }
+    };
+    let a = match args.first() {
+        Some(s) => long_val(s)?,
+        None => return Err(VmError::NullPointerException),
+    };
+    let b = match args.get(1) {
+        Some(s) => long_val(s)?,
+        None => return Err(VmError::NullPointerException),
+    };
+    Ok(Some(Slot::Int(ordering_to_int(a.cmp(&b)))))
+}
+
+// ---- Double class natives ----
+
+/// Native: `Double.parseDouble(String)` — parses string to double.
+pub(crate) fn native_double_parsedouble(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let str_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Reference",
+                got: "other",
+            });
+        }
+    };
+    let s = heap.get(str_ref)?.string_value.clone().unwrap_or_default();
+    let val: f64 = s.trim().parse().map_err(|_| VmError::JavaException {
+        class_name: "java/lang/NumberFormatException".to_string(),
+    })?;
+    Ok(Some(Slot::Double(val)))
+}
+
+/// Native: `Double.valueOf(double)` — boxes double into Double object.
+pub(crate) fn native_double_valueof(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let val = match args.first() {
+        Some(Slot::Double(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Double",
+                got: "other",
+            });
+        }
+    };
+    let r = heap.allocate("java/lang/Double".to_string(), 1);
+    heap.get_mut(r).unwrap().fields[0] = Slot::Double(val);
+    Ok(Some(Slot::Reference(Some(r))))
+}
+
+/// Native: `Double.doubleValue()` — unboxes Double to double.
+pub(crate) fn native_double_doublevalue(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let val = heap.get(this_ref)?.fields[0];
+    Ok(Some(val))
+}
+
+// ---- Float class native ----
+
+/// Native: `Float.parseFloat(String)` — parses string to float.
+pub(crate) fn native_float_parsefloat(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let str_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        Some(Slot::Reference(None)) => return Err(VmError::NullPointerException),
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Reference",
+                got: "other",
+            });
+        }
+    };
+    let s = heap.get(str_ref)?.string_value.clone().unwrap_or_default();
+    let val: f32 = s.trim().parse().map_err(|_| VmError::JavaException {
+        class_name: "java/lang/NumberFormatException".to_string(),
+    })?;
+    Ok(Some(Slot::Float(val)))
+}
+
+// ---- Boolean class native ----
+
+/// Native: `Boolean.parseBoolean(String)` — case-insensitive "true" → 1, else 0.
+pub(crate) fn native_boolean_parseboolean(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    match args.first() {
+        Some(Slot::Reference(Some(r))) => {
+            let s = heap.get(*r)?.string_value.clone().unwrap_or_default();
+            let val = s.eq_ignore_ascii_case("true");
+            Ok(Some(Slot::Int(i32::from(val))))
+        }
+        Some(Slot::Reference(None)) => Ok(Some(Slot::Int(0))),
+        _ => Err(VmError::TypeMismatch {
+            expected: "Reference",
+            got: "other",
+        }),
+    }
+}
+
+
+/// Index of a named static field within `ctx.static_fields`.
+pub(crate) fn static_field_idx(ctx: &ClassContext, name: &str) -> VmResult<usize> {
+    ctx.fields
+        .iter()
+        .filter(|f| f.is_static)
+        .position(|f| f.name == name)
+        .ok_or(VmError::InvalidFieldref { index: 0 })
+}
+
+// ---------------------------------------------------------------------------
+// StringBuilder natives
+// ---------------------------------------------------------------------------
+
+/// Native: `StringBuilder.<init>()V` — initialise empty buffer.
+pub(crate) fn native_sb_init(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let obj = heap.get_mut(this_ref)?;
+    obj.string_value = Some(String::new());
+    Ok(None)
+}
+
+/// Native: `StringBuilder.<init>(Ljava/lang/String;)V` — init with string.
+pub(crate) fn native_sb_init_string(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let init_str = match args.get(1) {
+        Some(Slot::Reference(Some(r))) => heap.get(*r)?.string_value.clone().unwrap_or_default(),
+        _ => String::new(),
+    };
+    let obj = heap.get_mut(this_ref)?;
+    obj.string_value = Some(init_str);
+    Ok(None)
+}
+
+/// Native: `StringBuilder.append(Ljava/lang/String;)Ljava/lang/StringBuilder;`
+pub(crate) fn native_sb_append_string(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let append_str = match args.get(1) {
+        Some(Slot::Reference(Some(r))) => heap.get(*r)?.string_value.clone().unwrap_or_default(),
+        _ => "null".to_string(),
+    };
+    let obj = heap.get_mut(this_ref)?;
+    if let Some(ref mut buf) = obj.string_value {
+        buf.push_str(&append_str);
+    }
+    Ok(Some(Slot::Reference(Some(this_ref))))
+}
+
+/// Native: `StringBuilder.append(I)Ljava/lang/StringBuilder;`
+pub(crate) fn native_sb_append_int(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let val = match args.get(1) {
+        Some(Slot::Int(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int",
+                got: "other",
+            });
+        }
+    };
+    let obj = heap.get_mut(this_ref)?;
+    if let Some(ref mut buf) = obj.string_value {
+        buf.push_str(&val.to_string());
+    }
+    Ok(Some(Slot::Reference(Some(this_ref))))
+}
+
+/// Native: `StringBuilder.append(J)Ljava/lang/StringBuilder;`
+pub(crate) fn native_sb_append_long(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let val = match args.get(1) {
+        Some(Slot::Long(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Long",
+                got: "other",
+            });
+        }
+    };
+    let obj = heap.get_mut(this_ref)?;
+    if let Some(ref mut buf) = obj.string_value {
+        buf.push_str(&val.to_string());
+    }
+    Ok(Some(Slot::Reference(Some(this_ref))))
+}
+
+/// Native: `StringBuilder.append(D)Ljava/lang/StringBuilder;`
+pub(crate) fn native_sb_append_double(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let val = match args.get(1) {
+        Some(Slot::Double(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Double",
+                got: "other",
+            });
+        }
+    };
+    let obj = heap.get_mut(this_ref)?;
+    if let Some(ref mut buf) = obj.string_value {
+        buf.push_str(&val.to_string());
+    }
+    Ok(Some(Slot::Reference(Some(this_ref))))
+}
+
+/// Native: `StringBuilder.append(F)Ljava/lang/StringBuilder;`
+pub(crate) fn native_sb_append_float(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let val = match args.get(1) {
+        Some(Slot::Float(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Float",
+                got: "other",
+            });
+        }
+    };
+    let obj = heap.get_mut(this_ref)?;
+    if let Some(ref mut buf) = obj.string_value {
+        buf.push_str(&val.to_string());
+    }
+    Ok(Some(Slot::Reference(Some(this_ref))))
+}
+
+/// Native: `StringBuilder.append(Z)Ljava/lang/StringBuilder;`
+pub(crate) fn native_sb_append_boolean(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let val = match args.get(1) {
+        Some(Slot::Int(v)) => *v != 0,
+        _ => false,
+    };
+    let obj = heap.get_mut(this_ref)?;
+    if let Some(ref mut buf) = obj.string_value {
+        buf.push_str(if val { "true" } else { "false" });
+    }
+    Ok(Some(Slot::Reference(Some(this_ref))))
+}
+
+/// Native: `StringBuilder.append(C)Ljava/lang/StringBuilder;`
+pub(crate) fn native_sb_append_char(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let val = match args.get(1) {
+        Some(Slot::Int(v)) => char::from_u32((*v).cast_unsigned()).unwrap_or('\0'),
+        _ => '\0',
+    };
+    let obj = heap.get_mut(this_ref)?;
+    if let Some(ref mut buf) = obj.string_value {
+        buf.push(val);
+    }
+    Ok(Some(Slot::Reference(Some(this_ref))))
+}
+
+/// Native: `StringBuilder.toString()Ljava/lang/String;`
+pub(crate) fn native_sb_tostring(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let content = heap.get(this_ref)?.string_value.clone().unwrap_or_default();
+    let r = heap.allocate_string(content);
+    Ok(Some(Slot::Reference(Some(r))))
+}
+
+/// Native: `StringBuilder.length()I`
+pub(crate) fn native_sb_length(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let len = heap
+        .get(this_ref)?
+        .string_value
+        .as_ref()
+        .map_or(0, String::len);
+    Ok(Some(Slot::Int(i32::try_from(len).unwrap_or(i32::MAX))))
+}
+
+// Character natives
+// ---------------------------------------------------------------------------
+
+/// Helper: extract a `char` from a `Slot::Int` argument.
+fn slot_to_char(slot: &Slot) -> VmResult<char> {
+    match slot {
+        Slot::Int(v) => Ok(char::from_u32((*v).cast_unsigned()).unwrap_or('\0')),
+        _ => Err(VmError::TypeMismatch {
+            expected: "Int (char)",
+            got: "other",
+        }),
+    }
+}
+
+/// Native: `Character.isDigit(C)Z`
+pub(crate) fn native_char_is_digit(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let ch = slot_to_char(args.first().ok_or(VmError::StackUnderflow)?)?;
+    Ok(Some(Slot::Int(i32::from(ch.is_ascii_digit()))))
+}
+
+/// Native: `Character.isLetter(C)Z`
+pub(crate) fn native_char_is_letter(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let ch = slot_to_char(args.first().ok_or(VmError::StackUnderflow)?)?;
+    Ok(Some(Slot::Int(i32::from(ch.is_alphabetic()))))
+}
+
+/// Native: `Character.isWhitespace(C)Z`
+pub(crate) fn native_char_is_whitespace(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let ch = slot_to_char(args.first().ok_or(VmError::StackUnderflow)?)?;
+    Ok(Some(Slot::Int(i32::from(ch.is_whitespace()))))
+}
+
+/// Native: `Character.isUpperCase(C)Z`
+pub(crate) fn native_char_is_uppercase(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let ch = slot_to_char(args.first().ok_or(VmError::StackUnderflow)?)?;
+    Ok(Some(Slot::Int(i32::from(ch.is_uppercase()))))
+}
+
+/// Native: `Character.isLowerCase(C)Z`
+pub(crate) fn native_char_is_lowercase(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let ch = slot_to_char(args.first().ok_or(VmError::StackUnderflow)?)?;
+    Ok(Some(Slot::Int(i32::from(ch.is_lowercase()))))
+}
+
+/// Native: `Character.toUpperCase(C)C`
+pub(crate) fn native_char_to_uppercase(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let ch = slot_to_char(args.first().ok_or(VmError::StackUnderflow)?)?;
+    let upper = ch.to_uppercase().next().unwrap_or(ch);
+    Ok(Some(Slot::Int(upper as i32)))
+}
+
+/// Native: `Character.toLowerCase(C)C`
+pub(crate) fn native_char_to_lowercase(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let ch = slot_to_char(args.first().ok_or(VmError::StackUnderflow)?)?;
+    let lower = ch.to_lowercase().next().unwrap_or(ch);
+    Ok(Some(Slot::Int(lower as i32)))
+}
+
+/// Native: `Character.isLetterOrDigit(C)Z`
+pub(crate) fn native_char_is_letter_or_digit(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let ch = slot_to_char(args.first().ok_or(VmError::StackUnderflow)?)?;
+    Ok(Some(Slot::Int(i32::from(ch.is_alphanumeric()))))
+}
+
+/// Native: `Character.valueOf(C)Ljava/lang/Character;` — box a char.
+pub(crate) fn native_char_valueof(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let val = match args.first() {
+        Some(Slot::Int(v)) => *v,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int (char)",
+                got: "other",
+            });
+        }
+    };
+    let r = heap.allocate("java/lang/Character".to_string(), 1);
+    heap.get_mut(r).unwrap().fields[0] = Slot::Int(val);
+    Ok(Some(Slot::Reference(Some(r))))
+}
+
+/// Native: `Character.charValue()C` — unbox Character to char.
+pub(crate) fn native_char_charvalue(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let val = heap.get(this_ref)?.fields[0];
+    Ok(Some(val))
+}
+
+// ---------------------------------------------------------------------------
+// ArrayList natives
+// ---------------------------------------------------------------------------
+
+/// Native: `ArrayList.<init>()V` — initializes with size=0.
+pub(crate) fn native_arraylist_init(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    heap.get_mut(this_ref)?.fields[0] = Slot::Int(0);
+    Ok(None)
+}
+
+/// Native: `ArrayList.add(Object)Z` — appends element, returns true.
+pub(crate) fn native_arraylist_add(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let element = args.get(1).copied().unwrap_or(Slot::Reference(None));
+    let obj = heap.get_mut(this_ref)?;
+    match obj.fields.first_mut() {
+        Some(Slot::Int(sz)) => *sz += 1,
+        _ => return Err(VmError::NullPointerException), // shouldn't happen; init sets fields[0]=Int(0)
+    }
+    obj.fields.push(element);
+    Ok(Some(Slot::Int(1))) // boolean true
+}
+
+/// Native: `ArrayList.get(I)Object` — returns element at index.
+#[allow(clippy::cast_sign_loss)]
+pub(crate) fn native_arraylist_get(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let idx = match args.get(1) {
+        Some(Slot::Int(i)) => *i as usize,
+        _ => {
+            return Err(VmError::TypeMismatch {
+                expected: "Int",
+                got: "other",
+            });
+        }
+    };
+    let obj = heap.get(this_ref)?;
+    obj.fields.get(idx + 1).map_or_else(
+        || {
+            Err(VmError::JavaException {
+                class_name: "java/lang/ArrayIndexOutOfBoundsException".to_string(),
+            })
+        },
+        |slot| Ok(Some(*slot)),
+    )
+}
+
+/// Native: `ArrayList.size()I`
+pub(crate) fn native_arraylist_size(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let obj = heap.get(this_ref)?;
+    match obj.fields.first() {
+        Some(Slot::Int(sz)) => Ok(Some(Slot::Int(*sz))),
+        _ => Ok(Some(Slot::Int(0))),
+    }
+}
+
+/// Native: `ArrayList.iterator()Iterator` — creates an `ArrayListIterator`.
+pub(crate) fn native_arraylist_iterator(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let iter_ref = heap.allocate("duke/util/ArrayListIterator".to_string(), 2);
+    {
+        let iter_obj = heap.get_mut(iter_ref)?;
+        iter_obj.fields[0] = Slot::Reference(Some(this_ref));
+        iter_obj.fields[1] = Slot::Int(0);
+    }
+    Ok(Some(Slot::Reference(Some(iter_ref))))
+}
+
+/// Native: `ArrayList.sort(Comparator)V` — sorts in-place using insertion sort,
+/// calling `compareTo` on each element pair via the interpreter callback.
+///
+/// Only null Comparator (natural ordering via `compareTo`) is supported.
+pub(crate) fn array_list_sort(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    output: &mut dyn Write,
+    _control: &mut NativeControl,
+    invoke: &mut InvokeFn<'_>,
+) -> VmResult<Option<Slot>> {
+    // args[0] = ArrayList ref, args[1] = Comparator (null = natural ordering)
+    let list_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+
+    // Fix 1: use the correct error variant for unsupported non-null Comparator.
+    if !matches!(args.get(1), Some(Slot::Reference(None)) | None) {
+        return Err(VmError::Unimplemented {
+            mnemonic: "ArrayList.sort(non-null Comparator)",
+        });
+    }
+
+    // Fix 2: guard against a negative size stored in fields[0].
+    let size = match heap.get(list_ref)?.fields.first() {
+        Some(Slot::Int(n)) if *n >= 0 => usize::try_from(*n).unwrap_or(0),
+        Some(Slot::Int(n)) => return Err(VmError::NegativeArraySize { size: *n }),
+        _ => return Ok(None),
+    };
+
+    if size <= 1 {
+        return Ok(None);
+    }
+
+    // Collect element refs (fields[1..=size]).
+    let mut elems: Vec<u64> = (1..=size)
+        .filter_map(|i| match heap.get(list_ref).ok()?.fields.get(i) {
+            Some(Slot::Reference(Some(r))) => Some(*r),
+            _ => None,
+        })
+        .collect();
+
+    // Fix 3: malformed list → InvalidRef, not silent Ok(None).
+    if elems.len() != size {
+        return Err(VmError::InvalidRef { address: list_ref });
+    }
+
+    // Insertion sort — O(n²), correct, easy to verify.
+    for i in 1..elems.len() {
+        let mut key = elems[i];
+        let mut j = i;
+        while j > 0 {
+            let receiver = elems[j - 1];
+            let class_name = heap.get(receiver)?.class_name.clone();
+            let cmp = invoke(
+                heap,
+                output,
+                &class_name,
+                COMPARE_TO_METHOD,
+                COMPARE_TO_OBJECT_DESC,
+                vec![Slot::Reference(Some(receiver)), Slot::Reference(Some(key))],
+            )?;
+
+            // Patch stale young-gen refs if a minor GC fired during the callback.
+            // Guarded by `has_pending_forwards` so the common (no-GC) path pays
+            // only one bool check instead of O(n) HashMap probes.
+            if heap.has_pending_forwards() {
+                for elem in &mut elems {
+                    let mut slot = Slot::Reference(Some(*elem));
+                    heap.apply_forward(&mut slot);
+                    if let Slot::Reference(Some(r)) = slot {
+                        *elem = r;
+                    }
+                }
+                // Re-read key after forwarding patch (it lives outside elems
+                // during the innermost loop iteration).
+                let mut key_slot = Slot::Reference(Some(key));
+                heap.apply_forward(&mut key_slot);
+                if let Slot::Reference(Some(r)) = key_slot {
+                    key = r;
+                }
+            }
+
+            // Fix 4: explicit error on non-Int compareTo return.
+            match cmp {
+                Some(Slot::Int(n)) if n <= 0 => break,
+                Some(Slot::Int(_)) => {} // n > 0, keep shifting
+                _ => {
+                    return Err(VmError::TypeMismatch {
+                        expected: "Int",
+                        got: "other",
+                    });
+                }
+            }
+            elems[j] = elems[j - 1];
+            j -= 1;
+        }
+        elems[j] = key;
+    }
+
+    // Write sorted elements back using the write barrier.
+    for (i, &r) in elems.iter().enumerate() {
+        heap.write_field(list_ref, i + 1, Slot::Reference(Some(r)))?;
+    }
+
+    Ok(None)
+}
+
+// ---------------------------------------------------------------------------
+// Collections natives
+// ---------------------------------------------------------------------------
+
+/// Native: `Collections.sort(List)V` — delegates to the list's sort(null) method.
+///
+/// `Collections.sort(list)` is compiled by javac as
+/// `invokestatic java/util/Collections.sort:(Ljava/util/List;)V`.
+/// We forward to the runtime class's `sort(Comparator=null)`, which for an
+/// `ArrayList` performs the insertion-sort-with-compareTo callback.
+pub(crate) fn native_collections_sort(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    output: &mut dyn Write,
+    _control: &mut NativeControl,
+    invoke: &mut InvokeFn<'_>,
+) -> VmResult<Option<Slot>> {
+    // args[0] = List ref
+    let list_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    // Dispatch on the actual runtime class so any List implementation works.
+    let class_name = heap.get(list_ref)?.class_name.clone();
+    invoke(
+        heap,
+        output,
+        &class_name,
+        "sort",
+        SORT_COMPARATOR_DESC,
+        vec![Slot::Reference(Some(list_ref)), Slot::Reference(None)],
+    )?;
+    Ok(None)
+}
+
+// ---------------------------------------------------------------------------
+// ArrayListIterator natives
+// ---------------------------------------------------------------------------
+
+/// Native: `ArrayListIterator.<init>` — no-op; fields set directly by `native_arraylist_iterator`.
+#[allow(clippy::unnecessary_wraps)]
+pub(crate) fn native_arraylist_iter_init(
+    _args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    Ok(None)
+}
+
+/// Native: `ArrayListIterator.hasNext()Z`
+pub(crate) fn native_arraylist_iter_hasnext(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let iter_obj = heap.get(this_ref)?;
+    let list_ref = match iter_obj.fields.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Ok(Some(Slot::Int(0))),
+    };
+    let cursor = match iter_obj.fields.get(1) {
+        Some(Slot::Int(i)) => *i,
+        _ => 0,
+    };
+    let list_size = match heap.get(list_ref)?.fields.first() {
+        Some(Slot::Int(sz)) => *sz,
+        _ => 0,
+    };
+    Ok(Some(Slot::Int(i32::from(cursor < list_size))))
+}
+
+/// Native: `ArrayListIterator.next()Object` — returns element at cursor, advances cursor.
+#[allow(clippy::cast_sign_loss)]
+pub(crate) fn native_arraylist_iter_next(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let (list_ref, cursor) = {
+        let iter_obj = heap.get(this_ref)?;
+        let lr = match iter_obj.fields.first() {
+            Some(Slot::Reference(Some(r))) => *r,
+            _ => return Err(VmError::NullPointerException),
+        };
+        let c = match iter_obj.fields.get(1) {
+            Some(Slot::Int(i)) => *i,
+            _ => 0,
+        };
+        (lr, c)
+    };
+    let element = {
+        let list_obj = heap.get(list_ref)?;
+        match list_obj.fields.get(cursor as usize + 1) {
+            Some(slot) => *slot,
+            None => {
+                return Err(VmError::JavaException {
+                    class_name: "java/util/NoSuchElementException".to_string(),
+                });
+            }
+        }
+    };
+    heap.get_mut(this_ref)?.fields[1] = Slot::Int(cursor + 1);
+    Ok(Some(element))
+}
+
+// ---- Double.isNaN ----
+
+/// Native: `Double.isNaN(D)Z` — returns 1 if value is NaN.
+#[allow(clippy::unnecessary_wraps)]
+pub(crate) fn native_double_isnan(
+    args: &[Slot],
+    _heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    match args.first() {
+        Some(Slot::Double(v)) => Ok(Some(Slot::Int(i32::from(v.is_nan())))),
+        _ => Ok(Some(Slot::Int(0))),
+    }
+}
+
+/// Native: `Double.compareTo(Object)` — compares two boxed Doubles.
+pub(crate) fn native_double_compareto(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let double_val = |s: &Slot| -> VmResult<f64> {
+        match s {
+            Slot::Reference(Some(r)) => match heap.get(*r)?.fields.first() {
+                Some(Slot::Double(n)) => Ok(*n),
+                _ => Err(VmError::InvalidRef { address: *r }),
+            },
+            _ => Err(VmError::NullPointerException),
+        }
+    };
+    let a = match args.first() {
+        Some(s) => double_val(s)?,
+        None => return Err(VmError::NullPointerException),
+    };
+    let b = match args.get(1) {
+        Some(s) => double_val(s)?,
+        None => return Err(VmError::NullPointerException),
+    };
+    // Use total_cmp: implements Java's total order where NaN > +∞ > … > -∞.
+    Ok(Some(Slot::Int(ordering_to_int(a.total_cmp(&b)))))
+}
+
+// ---- Arrays natives ----
+
+/// Native: `Arrays.fill(int[], int)` — fills all elements with val.
+pub(crate) fn native_arrays_fill_int(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let arr_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let val = match args.get(1) {
+        Some(Slot::Int(v)) => Slot::Int(*v),
+        _ => Slot::Int(0),
+    };
+    let obj = heap.get_mut(arr_ref)?;
+    for slot in &mut obj.fields {
+        *slot = val;
+    }
+    Ok(None)
+}
+
+/// Native: `Arrays.fill(Object[], Object)` — fills all elements with val.
+pub(crate) fn native_arrays_fill_object(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let arr_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let val = args.get(1).copied().unwrap_or(Slot::Reference(None));
+    let obj = heap.get_mut(arr_ref)?;
+    for slot in &mut obj.fields {
+        *slot = val;
+    }
+    Ok(None)
+}
+
+/// Native: `Arrays.copyOf(int[], int)` — copies to new int[] of given length.
+pub(crate) fn native_arrays_copyof_int(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let src_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let new_len = match args.get(1) {
+        Some(Slot::Int(n)) if *n >= 0 => usize::try_from(*n).unwrap_or(0),
+        Some(Slot::Int(n)) => return Err(VmError::NegativeArraySize { size: *n }),
+        _ => 0,
+    };
+    let src_fields = heap.get(src_ref)?.fields.clone();
+    let dst_ref = heap.allocate("[I".to_string(), new_len);
+    let dst = heap.get_mut(dst_ref)?;
+    for i in 0..new_len {
+        dst.fields[i] = src_fields.get(i).copied().unwrap_or(Slot::Int(0));
+    }
+    Ok(Some(Slot::Reference(Some(dst_ref))))
+}
+
+/// Native: `Arrays.copyOf(Object[], int)` — copies to new Object[] of given length.
+pub(crate) fn native_arrays_copyof_object(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let src_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let new_len = match args.get(1) {
+        Some(Slot::Int(n)) if *n >= 0 => usize::try_from(*n).unwrap_or(0),
+        Some(Slot::Int(n)) => return Err(VmError::NegativeArraySize { size: *n }),
+        _ => 0,
+    };
+    let src_fields = heap.get(src_ref)?.fields.clone();
+    let dst_ref = heap.allocate("[Ljava/lang/Object;".to_string(), new_len);
+    let dst = heap.get_mut(dst_ref)?;
+    for i in 0..new_len {
+        dst.fields[i] = src_fields.get(i).copied().unwrap_or(Slot::Reference(None));
+    }
+    Ok(Some(Slot::Reference(Some(dst_ref))))
+}
+
+/// Native: `Arrays.sort(int[])` — sorts fields in place.
+pub(crate) fn native_arrays_sort_int(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let arr_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let obj = heap.get_mut(arr_ref)?;
+    obj.fields.sort_by(|a, b| match (a, b) {
+        (Slot::Int(x), Slot::Int(y)) => x.cmp(y),
+        _ => std::cmp::Ordering::Equal,
+    });
+    Ok(None)
+}
+
+// ---------------------------------------------------------------------------
+// HashMap natives
+// ---------------------------------------------------------------------------
+
+/// Semantic equality for `HashMap` keys: compares by `string_value` for heap strings,
+/// or by the first field (e.g. intValue) for boxed numerics, or by reference identity.
+pub(crate) fn slots_equal(a: &Slot, b: &Slot, heap: &duke_gc::Heap) -> bool {
+    match (a, b) {
+        (Slot::Reference(None), Slot::Reference(None)) => true,
+        (Slot::Reference(Some(ra)), Slot::Reference(Some(rb))) => {
+            if ra == rb {
+                return true;
+            }
+            let Ok(oa) = heap.get(*ra) else { return false };
+            let Ok(ob) = heap.get(*rb) else { return false };
+            if oa.class_name == "java/lang/String" || ob.class_name == "java/lang/String" {
+                return oa.string_value == ob.string_value;
+            }
+            if oa.class_name == ob.class_name {
+                oa.fields.first() == ob.fields.first()
+            } else {
+                false
+            }
+        }
+        _ => false,
+    }
+}
+
+/// Native: `HashMap.<init>()V` — initialises size counter at fields\[0\] to 0.
+fn find_hashmap_entry_index(fields: &[Slot], key: &Slot, heap: &duke_gc::Heap) -> Option<usize> {
+    (1..fields.len())
+        .step_by(2)
+        .find(|&i| i + 1 < fields.len() && slots_equal(&fields[i], key, heap))
+}
+
+pub(crate) fn native_hashmap_init(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let obj = heap.get_mut(this_ref)?;
+    if obj.fields.is_empty() {
+        obj.fields.push(Slot::Int(0));
+    } else {
+        obj.fields[0] = Slot::Int(0);
+    }
+    Ok(None)
+}
+
+/// Native: `HashMap.put(Object, Object)Object` — inserts or updates a key-value pair.
+/// Returns the old value if the key was already present, or null if it is new.
+pub(crate) fn native_hashmap_put(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let key = args.get(1).copied().unwrap_or(Slot::Reference(None));
+    let val = args.get(2).copied().unwrap_or(Slot::Reference(None));
+    // Clone fields to release the immutable borrow before mutating.
+    let fields = heap.get(this_ref)?.fields.clone();
+
+    if let Some(i) = find_hashmap_entry_index(&fields, &key, heap) {
+        let old = fields[i + 1];
+        heap.get_mut(this_ref)?.fields[i + 1] = val;
+        return Ok(Some(old));
+    }
+
+    // New key — append pair and bump size.
+    let obj = heap.get_mut(this_ref)?;
+    match obj.fields.first_mut() {
+        Some(Slot::Int(sz)) => *sz += 1,
+        _ => return Err(VmError::NullPointerException),
+    }
+    obj.fields.push(key);
+    obj.fields.push(val);
+    Ok(Some(Slot::Reference(None)))
+}
+
+/// Native: `HashMap.get(Object)Object` — returns value for key, or null if absent.
+pub(crate) fn native_hashmap_get(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let key = args.get(1).copied().unwrap_or(Slot::Reference(None));
+    let fields = heap.get(this_ref)?.fields.clone();
+
+    Ok(Some(
+        find_hashmap_entry_index(&fields, &key, heap)
+            .map_or(Slot::Reference(None), |i| fields[i + 1]),
+    ))
+}
+
+/// Native: `HashMap.containsKey(Object)Z` — returns 1 if key present, 0 otherwise.
+pub(crate) fn native_hashmap_contains_key(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let key = args.get(1).copied().unwrap_or(Slot::Reference(None));
+    let fields = heap.get(this_ref)?.fields.clone();
+
+    if find_hashmap_entry_index(&fields, &key, heap).is_some() {
+        Ok(Some(Slot::Int(1)))
+    } else {
+        Ok(Some(Slot::Int(0)))
+    }
+}
+
+/// Native: `HashMap.size()I` — returns entry count from fields\[0\].
+pub(crate) fn native_hashmap_size(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => Ok(Some(Slot::Int(*n))),
+        _ => Ok(Some(Slot::Int(0))),
+    }
+}
+
+/// Native: `HashMap.remove(Object)Object` — removes a key-value pair, returns old value or null.
+/// Uses swap-remove (swaps target pair with last pair) for O(1) deletion.
+pub(crate) fn native_hashmap_remove(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let key = args.get(1).copied().unwrap_or(Slot::Reference(None));
+    let fields = heap.get(this_ref)?.fields.clone();
+
+    if let Some(i) = find_hashmap_entry_index(&fields, &key, heap) {
+        let old_val = fields[i + 1];
+        let obj = heap.get_mut(this_ref)?;
+        let last_val_idx = obj.fields.len() - 1;
+        let last_key_idx = obj.fields.len() - 2;
+        obj.fields.swap(i + 1, last_val_idx);
+        obj.fields.swap(i, last_key_idx);
+        obj.fields.truncate(obj.fields.len() - 2);
+        match obj.fields.first_mut() {
+            Some(Slot::Int(sz)) => *sz -= 1,
+            _ => return Err(VmError::NullPointerException),
+        }
+        Ok(Some(old_val))
+    } else {
+        Ok(Some(Slot::Reference(None)))
+    }
+}
+
+/// Native: `HashMap.isEmpty()Z` — returns 1 if size == 0, else 0.
+pub(crate) fn native_hashmap_is_empty(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(0)) => Ok(Some(Slot::Int(1))),
+        Some(Slot::Int(_)) => Ok(Some(Slot::Int(0))),
+        _ => Ok(Some(Slot::Int(1))),
+    }
+}
+
+/// Native: `HashMap.getOrDefault(Object, Object)Object` — returns value for key, or default if absent.
+pub(crate) fn native_hashmap_get_or_default(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let key = args.get(1).copied().unwrap_or(Slot::Reference(None));
+    let default = args.get(2).copied().unwrap_or(Slot::Reference(None));
+    let fields = heap.get(this_ref)?.fields.clone();
+
+    Ok(Some(
+        find_hashmap_entry_index(&fields, &key, heap).map_or(default, |i| fields[i + 1]),
+    ))
+}
+
+// ---------------------------------------------------------------------------
+// HashSet natives
+// ---------------------------------------------------------------------------
+
+/// Native: `HashSet.<init>()V` — initialises size counter at fields\[0\] to 0.
+fn find_hashset_entry_index(
+    fields: &[Slot],
+    element: &Slot,
+    heap: &duke_gc::Heap,
+) -> Option<usize> {
+    fields
+        .iter()
+        .skip(1)
+        .position(|field| slots_equal(field, element, heap))
+        .map(|idx| idx + 1)
+}
+
+pub(crate) fn native_hashset_init(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let obj = heap.get_mut(this_ref)?;
+    if obj.fields.is_empty() {
+        obj.fields.push(Slot::Int(0));
+    } else {
+        obj.fields[0] = Slot::Int(0);
+    }
+    Ok(None)
+}
+
+/// Native: `HashSet.add(Object)Z` — adds element if not already present.
+/// Returns 1 if added, 0 if element was already in the set.
+pub(crate) fn native_hashset_add(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let element = args.get(1).copied().unwrap_or(Slot::Reference(None));
+    let fields = heap.get(this_ref)?.fields.clone();
+    // fields[0] = size, fields[1..] = elements
+    if find_hashset_entry_index(&fields, &element, heap).is_some() {
+        return Ok(Some(Slot::Int(0))); // duplicate
+    }
+
+    let obj = heap.get_mut(this_ref)?;
+    match obj.fields.first_mut() {
+        Some(Slot::Int(sz)) => *sz += 1,
+        _ => return Err(VmError::NullPointerException),
+    }
+    obj.fields.push(element);
+    Ok(Some(Slot::Int(1)))
+}
+
+/// Native: `HashSet.contains(Object)Z` — returns 1 if element is present, 0 otherwise.
+pub(crate) fn native_hashset_contains(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let element = args.get(1).copied().unwrap_or(Slot::Reference(None));
+    let fields = heap.get(this_ref)?.fields.clone();
+
+    if find_hashset_entry_index(&fields, &element, heap).is_some() {
+        Ok(Some(Slot::Int(1)))
+    } else {
+        Ok(Some(Slot::Int(0)))
+    }
+}
+
+/// Native: `HashSet.remove(Object)Z` — removes element if present, returns 1 if removed, 0 if absent.
+/// Uses swap-remove (swaps target with last element) for O(1) deletion.
+pub(crate) fn native_hashset_remove(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    let element = args.get(1).copied().unwrap_or(Slot::Reference(None));
+    let fields = heap.get(this_ref)?.fields.clone();
+
+    if let Some(i) = find_hashset_entry_index(&fields, &element, heap) {
+        let obj = heap.get_mut(this_ref)?;
+        let last_idx = obj.fields.len() - 1;
+        obj.fields.swap(i, last_idx);
+        obj.fields.truncate(obj.fields.len() - 1);
+        match obj.fields.first_mut() {
+            Some(Slot::Int(sz)) => *sz -= 1,
+            _ => return Err(VmError::NullPointerException),
+        }
+        Ok(Some(Slot::Int(1)))
+    } else {
+        Ok(Some(Slot::Int(0)))
+    }
+}
+
+/// Native: `HashSet.size()I` — returns element count from fields\[0\].
+pub(crate) fn native_hashset_size(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(n)) => Ok(Some(Slot::Int(*n))),
+        _ => Ok(Some(Slot::Int(0))),
+    }
+}
+
+/// Native: `HashSet.isEmpty()Z` — returns 1 if size == 0, else 0.
+pub(crate) fn native_hashset_is_empty(
+    args: &[Slot],
+    heap: &mut duke_gc::Heap,
+    _out: &mut dyn Write,
+    _control: &mut NativeControl,
+) -> VmResult<Option<Slot>> {
+    let this_ref = match args.first() {
+        Some(Slot::Reference(Some(r))) => *r,
+        _ => return Err(VmError::NullPointerException),
+    };
+    match heap.get(this_ref)?.fields.first() {
+        Some(Slot::Int(0)) => Ok(Some(Slot::Int(1))),
+        Some(Slot::Int(_)) => Ok(Some(Slot::Int(0))),
+        _ => Ok(Some(Slot::Int(1))),
+    }
+}
+
+// ---------------------------------------------------------------------------
+// GC root gathering
+// ---------------------------------------------------------------------------
